% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_group_info.R
\name{annotate_group_info}
\alias{annotate_group_info}
\title{Annotate an adxmiture barplot with information about the populations}
\usage{
annotate_group_info(group)
}
\arguments{
\item{group}{a vector with the names of the populations for each
individuals. Note that individuals belonging to a given populaiton need to
be in a block, all adjacent to each other}
}
\value{
modifier for a ggplot, added with the usual '+'
}
\description{
Add vertical lines to separate individuals from different populations,
and add optional x labels with their names
}
