% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyverse.R
\name{arrange.gen_tbl}
\alias{arrange.gen_tbl}
\title{An arrange method for \code{gen_tibble} objects}
\usage{
\method{arrange}{gen_tbl}(..., deparse.level = 1)
}
\arguments{
\item{...}{a gen_tibble and a data.frame or tibble}

\item{deparse.level}{an integer controlling the construction of column names.}
}
\value{
a \code{gen_tibble}
}
\description{
An arrange method for \code{gen_tibble} objects
}
\examples{
test_gt <- load_example_gt("gen_tbl")
test_gt \%>\% arrange(id)
}
