% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_gen_tibble.R
\name{as_gen_tibble}
\alias{as_gen_tibble}
\alias{as_gen_tibble.genlight}
\title{Cast to a \code{gen_tibble}}
\usage{
as_gen_tibble(.x, ...)

\method{as_gen_tibble}{genlight}(.x, ..., ignore_null_slots = TRUE)
}
\arguments{
\item{.x}{the object to convert}

\item{...}{additional arguments to pass to a given method}

\item{ignore_null_slots}{boolean on whether null slots should be ignored
(and thus filled with default values), or whether an error will be thrown}
}
\value{
a \code{gen_tbl} object
}
\description{
Convert an object to a \code{gen_tibble}
}
