% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_genlight.R
\name{as_genlight}
\alias{as_genlight}
\title{Cast a \code{gen_tibble} to a \code{genlight} object}
\usage{
as_genlight(x, ...)
}
\arguments{
\item{x}{the object to convert}

\item{...}{additional arguments (currently unused)}
}
\value{
a \code{genlight} object
}
\description{
Convert an \link{gen_tibble} object to a \code{\link[adegenet:genlight]{adegenet::genlight}} object
}
