% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q_matrix.R
\name{as_q_matrix}
\alias{as_q_matrix}
\title{Convert a standard matrix to a \code{q_matrix} object}
\usage{
as_q_matrix(x)
}
\arguments{
\item{x}{A matrix or a data frame}
}
\value{
A \code{q_matrix} object
}
\description{
Takes a single Q matrix that exists as either a matrix or a data frame and
returns a \code{q_matrix} object.
}
\keyword{internal}
