% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_q_matrix.R
\name{autoplot.q_matrix}
\alias{autoplot.q_matrix}
\title{Autoplots for \code{q_matrix} objects}
\usage{
\method{autoplot}{q_matrix}(object, data = NULL, annotate_group = TRUE, ...)
}
\arguments{
\item{object}{A Q matrix object (as returned by \code{\link[=as_q_matrix]{as_q_matrix()}}).}

\item{data}{An associated tibble (e.g. a \code{\link{gen_tibble}}), with the individuals in the same order as the data used to
generate the Q matrix}

\item{annotate_group}{Boolean determining whether to annotate the plot with the
group information}

\item{...}{not currently used.}
}
\value{
a barplot of individuals, coloured by ancestry proportion
}
\description{
Autoplots for \code{q_matrix} objects
}
