% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_report_loci.R
\name{autoplot.qc_report_loci}
\alias{autoplot.qc_report_loci}
\title{Autoplots for \code{qc_report_loci} objects}
\usage{
\method{autoplot}{qc_report_loci}(
  object,
  type = c("overview", "all", "missing", "missing low maf", "missing high maf", "maf",
    "hwe", "significant hwe"),
  maf_threshold = 0.05,
  miss_threshold = 0.01,
  hwe_p = 0.01,
  ...
)
}
\arguments{
\item{object}{an object of class \code{qc_report_loci}}

\item{type}{the type of plot (one of \code{overview}, \code{all}, \code{missing}, \verb{missing low maf}, \verb{missing high maf}, \code{maf}, \code{hwe}, and \verb{significant hwe})}

\item{maf_threshold}{default 0.5, a threshold for the accepted rate of minor
allele frequency of loci}

\item{miss_threshold}{default 0.01, a threshold for the accepted rate of
missingness per loci}

\item{hwe_p}{default 0.01, a threshold of significance for Hardy-Weinberg
exact p-values}

\item{...}{not currently used.}
}
\value{
a \code{ggplot2} object
}
\description{
For \code{qc_report_loci}, the following types of plots are available:
\itemize{
\item \code{overview}: an UpSet plot, giving counts of snps over the threshold for
missingness, minor allele frequency, and Hardy-Weinberg equilibrium P-value,
and visualising the interaction between these
\item \code{all}: a four panel plot, containing \verb{missing high maf}, \verb{missing low maf},
\code{hwe}, and \verb{significant hwe} plots
\item \code{missing}: a histogram of proportion of missing data
\item \verb{missing low maf}: a histogram of the proportion of missing data for
snps with low minor allele frequency
\item \verb{missing high maf}:a histogram of the proportion of missing data for
snps with high minor allele frequency
\item \code{maf}: a histogram of minor allele frequency
\item \code{hwe}: a histogram of HWE exact test p-values
\item \verb{significant hwe}: a histogram of significant HWE exact test p-values
}
}
\details{
\code{autoplot} produces simple plots to quickly inspect an object. They are not
customisable; we recommend that you use \code{ggplot2} to produce publication
ready plots.
}
