% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_report_loci.R
\name{autoplot_l_qc_missing}
\alias{autoplot_l_qc_missing}
\title{Internal missingness plot function}
\usage{
autoplot_l_qc_missing(
  object,
  miss_threshold,
  maf_low_thresh = NULL,
  maf_high_thresh = NULL,
  ...
)
}
\arguments{
\item{object}{an object of class \code{qc_report_loci}}

\item{miss_threshold}{a threshold for the accepted rate of missingness per
loci}

\item{maf_low_thresh}{a threshold for the accepted rate of minor allele
frequency of loci}

\item{maf_high_thresh}{a threshold for the accepted rate of minor allele
frequency of loci}

\item{...}{not currently used.}
}
\value{
a \code{ggplot2} object
}
\description{
Internal missingness plot function
}
\keyword{internal}
