% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_tibble.R
\name{bignsnp_from_dfs}
\alias{bignsnp_from_dfs}
\title{Create a bigSNP object from data.frames}
\usage{
bignsnp_from_dfs(genotypes, indiv_meta, loci)
}
\arguments{
\item{genotypes}{a genotype matrix}

\item{indiv_meta}{the individual meta information}

\item{loci}{the loci table}
}
\description{
This function expects the indiv_meta and loci to have the correct columns
}
\keyword{internal}
