% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_admix_methods.R
\name{c.gt_admix}
\alias{c.gt_admix}
\title{Combine method for gt_admix objects}
\usage{
\method{c}{gt_admix}(...)
}
\arguments{
\item{...}{A list of \code{gt_admix} objects}
}
\value{
A \code{gt_admix} object with the combined data
}
\description{
Combine method for gt_admix objects
}
