% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyverse.R
\name{$<-.gen_tbl}
\alias{$<-.gen_tbl}
\title{A $ method for \code{gen_tibble} objects}
\usage{
\method{$}{gen_tbl}(x, i) <- value
}
\arguments{
\item{x}{a gen_tibble}

\item{i}{column name}

\item{value}{a value to assign}
}
\value{
a \code{gen_tibble}
}
\description{
A $ method for \code{gen_tibble} objects
}
\examples{
example_gt <- load_example_gt("gen_tbl")

# Add a new column
example_gt$region <- "East"

example_gt
}
