% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windos_stats_generic.R
\name{chrom_window_by_pos}
\alias{chrom_window_by_pos}
\title{Internal function to compute the windows over a chromosome by position}
\usage{
chrom_window_by_pos(
  x_sub,
  position_sub,
  i_chrom,
  operator_func,
  window_size,
  window_step,
  min_loci,
  complete
)
}
\arguments{
\item{x_sub}{A vector containing the per locus estimates for a single
chromosome.}

\item{i_chrom}{The chromosome label}

\item{operator_func}{The function to use for the window statistics. This
should be a function from the slider package, such as slide_mean or
slide_sum.}

\item{window_size}{The size of the window to use for the estimates.}

\item{window_step}{The step size to use for the windows.}

\item{min_loci}{The minimum number of loci required to calculate a window
statistic. If the number of loci in a window is less than this, the window
statistic will be NA.}

\item{complete}{Should the function be evaluated on complete windows only?
If FALSE, the default, then partial computations will be allowed at the end
of the chromosome.}
}
\value{
A dataframe with columns: 'chromosome', 'start', 'end', 'stats', and
'n_snps'. The 'stats' column contains the mean of the per locus estimates
in the window, and 'n_snps' contains the number of loci in the window.
}
\description{
Internal function to compute the windows over a chromosome by position
}
\keyword{internal}
