% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_int_names.R
\name{col.names.matrix_int_names}
\alias{col.names.matrix_int_names}
\title{Get column names (both integer and character)}
\usage{
\method{col.names}{matrix_int_names}(x)
}
\description{
Get column names (both integer and character)
}
