% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_int_names.R
\name{col.names}
\alias{col.names}
\alias{col.names<-}
\title{Get or set column names for matrix_int_names}
\usage{
col.names(x)

col.names(x) <- value
}
\arguments{
\item{x}{A matrix_int_names object}

\item{value}{a vector of integer or character column names to set, replacing
the current one.}
}
\value{
Integer or character vector of column names, or NULL
}
\description{
This function reliably returns integer names if present, otherwise character
dimnames. Use this instead of colnames() if you need guaranteed dispatch.
}
