% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_int_names.R
\name{col_names<-.matrix_int_names}
\alias{col_names<-.matrix_int_names}
\title{Set column names (integer or character)}
\usage{
\method{col_names}{matrix_int_names}(x) <- value
}
\description{
Set column names (integer or character)
}
