% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_loci.R
\name{count_loci}
\alias{count_loci}
\alias{count_loci.tbl_df}
\alias{count_loci.vctrs_bigSNP}
\title{Count the number of loci in a \code{gen_tibble}}
\usage{
count_loci(.x, ...)

\method{count_loci}{tbl_df}(.x, ...)

\method{count_loci}{vctrs_bigSNP}(.x, ...)
}
\arguments{
\item{.x}{a \code{\link{gen_tibble}}, or a vector of class \code{vctrs_bigSNP} (usually the
\code{genotype} column of a \code{\link{gen_tibble}} object).}

\item{...}{currently unused.}
}
\value{
the number of loci
}
\description{
Count the number of loci in \code{gen_tibble} (or directly from its \code{genotype}
column).
}
\examples{
example_gt <- load_example_gt("gen_tbl")

example_gt \%>\% count_loci()
}
