% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_vcf_variants.R
\name{count_vcf_individuals}
\alias{count_vcf_individuals}
\title{Counts the number of individuals in a vcf file}
\usage{
count_vcf_individuals(file)
}
\arguments{
\item{file}{name and path of vcf file (it can be compressed)}
}
\value{
the number of individuals (an integer)
}
\description{
Count the number of VCF individuals by first parsing the header, and then
getting the first line, then separating on tabs.
}
\author{
author Henrik Bengtsson for the original \code{countLines} in \code{R.utils};
Andrea Manica for the modified version focussed on vcf; Max Carter-Brown modified
the file above.
}
\keyword{internal}
