% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_vcf_variants.R
\name{count_vcf_variants}
\alias{count_vcf_variants}
\title{Counts the number of variants in a vcf file}
\usage{
count_vcf_variants(file, chunk_size = 5e+07)
}
\arguments{
\item{file}{name and path of vcf file (it can be compressed)}

\item{chunk_size}{The number of bytes read in each chunk.}
}
\value{
the number of variants (an integer)
}
\description{
Count the number of VCF variants by first parsing the header, and then
rapidly counting the number of platform-independent
newlines (CR, LF, and CR+LF), including a last line with neither.
}
\examples{
vcf_path <- system.file("/extdata/anolis/punctatus_t70_s10_n46_filtered.recode.vcf.gz",
  package = "tidypopgen"
)
count_vcf_variants(vcf_path)
}
\author{
author Henrik Bengtsson for the original \code{countLines} in \code{R.utils};
Andrea Manica for the modified version focussed on vcf
}
\keyword{internal}
