% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base62_encode.R
\name{decode62}
\alias{decode62}
\title{Function to decode a base62 encoded string back into a pair of integers.}
\usage{
decode62(encoded_str, max_chr = NULL, num_digits_chr = NULL)
}
\arguments{
\item{encoded_str}{A base62 encoded string representing the combined
chromosome and position}

\item{max_chr}{Integer representing the maximum chromosome number expected.
This is used to determine the fixed width for decoding the chromosome
number. If left NULL, an error will be raised as this information is needed
for decoding.}

\item{num_digits_chr}{(Optional) Integer representing the number of
characters used to encode the chromosome number. If provided, this will
override the calculation based on \code{max_chr}.}
}
\value{
A data.frame with two columns: \code{chr} and \code{pos}, representing the
decoded chromosome numbers and positions.
}
\description{
This function reverses the encoding performed by \code{encode_62}, taking a base62
encoded string and extracting the original chromosome number and position. It
uses the fixed width determined by the maximum chromosome number to correctly
separate the two components of the encoding.
}
\examples{
encoded_coords <- encode62(c(1, 10, 260), c(1, 1000, 1000000))
print(encoded_coords)
decoded_coords <- decode62(encoded_coords, max_chr = 260)
print(decoded_coords)
}
