% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_docs.R
\docType{data}
\name{distruct_colours}
\alias{distruct_colours}
\title{Distruct colours}
\format{
A vector of 60 hex colours
}
\usage{
distruct_colours
}
\description{
Colours in the palette used by distruct
}
\keyword{datasets}
