% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotypes_dot_prod.R
\name{.genotypes_dot_prod}
\alias{.genotypes_dot_prod}
\title{Compute the the dot products of loci}
\usage{
.genotypes_dot_prod(
  .x,
  center = FALSE,
  scale = FALSE,
  alleles_as_units = FALSE,
  parallel = FALSE,
  n.cores = NULL
)
}
\arguments{
\item{.x}{a list of \code{\link[adegenet:SNPbin]{adegenet::SNPbin}} objects (usually the \code{genotype} column of
a \code{\link{gen_tibble}} object),}

\item{alleles_as_units}{a logical indicating whether alleles are considered
as units (i.e., a diploid genotype equals two samples, a triploid, three,
etc.) or whether individuals are considered as units of information.}
}
\value{
a vector of counts of NAs
}
\description{
This function compute the dot products between centere/scaled vectors of
loci. It is not
meant to be used as a general summary, but it is used within other functions to
compute intermediate quantities.
When ploidy varies across individuals, the outputs of this function depend
on whether the information units are individuals, or
alleles within individuals (see details for \code{\link[=.genotypes_sums]{.genotypes_sums()}}).
}
\details{
This function is a modified version
of \link[adegenet:glAux]{adegenet::glDotProd}, recoded to work on lists of \code{SNPbin} objects as used
in the \code{genotypes} column of \link{gen_tibble}.
}
\author{
Thibaut Jombart for the original \link[adegenet:glAux]{adegenet::glDotProd}, modified
by Andrea Manica for 'tidypopgen'
}
