% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyverse.R
\name{dplyr_reconstruct.grouped_gen_tbl}
\alias{dplyr_reconstruct.grouped_gen_tbl}
\title{A dplyr_reconstruct method for grouped \code{gen_tibble} objects}
\usage{
\method{dplyr_reconstruct}{grouped_gen_tbl}(data, template)
}
\arguments{
\item{data}{a gen_tibble and a data.frame or tibble}

\item{template}{a data.frame or tibble}
}
\value{
a grouped \code{gen_tibble}
}
\description{
A dplyr_reconstruct method for grouped \code{gen_tibble} objects
}
