% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base62_encode.R
\name{encode62}
\alias{encode62}
\title{Function to encode a pair of integers into a base62 string.}
\usage{
encode62(chr, pos, max_chr = NULL)
}
\arguments{
\item{chr}{Chromosomes, either character/factor, which will be cast to
integer, or integer representing the chromosome number.}

\item{pos}{Integer representing the position on the chromosome.}

\item{max_chr}{Integer representing the maximum chromosome number expected.
This is used to determine the fixed width for encoding the chromosome
number. If left NULL, the maximum value of \code{chr} will be used.}
}
\value{
A base62 encoded string representing the combined chromosome and
position.
}
\description{
This function is used to encode the chromosome and the coordinates of a SNP
into a compact string representation using base62 encoding. The first integer
represents the chromosome number, and the second integer represents the
position on that chromosome. Th e encoded string can be used as a unique
identifier for the SNP.
}
\details{
The two numbers are combined by exploiting the fact that the the number of
chromosome is usually small and known in advance (e.g., 24 for humans). This
allows the chromosome number to be encoded using a fixed number of base-62
digits, while the position can use a variable number of digits. This ensures
that the combined encoding is unambiguous and compact. Specifically, both
numbers are converted to base-62 using only ASCII characters. The position is
encoded first, using as many base-62 digits as needed. The first number is
then encoded using a fixed number of base-62 digits (chosen in advance to be
large enough to represent its maximum possible value) and left-padded with
zeros if necessary. The final string is formed by concatenating these two
encodings, with the fixed-width encoding of the first number placed at the
end. During decoding, the last fixed number of characters are read back as
the chromosome number, and the remaining prefix is decoded as the position,
allowing the original pair to be recovered unambiguously without any
separator characters.
}
\examples{
encoded_coords <- encode62(c(1, 10, 260), c(1, 1000, 1000000))
print(encoded_coords)
decoded_coords <- decode62(encoded_coords, max_chr = 260)
print(decoded_coords)
}
