% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_gen_tibble.R
\name{example_gen_tibble}
\alias{example_gen_tibble}
\title{Create example gen_tibble}
\usage{
example_gen_tibble()
}
\value{
an example object of the class \code{gen_tbl}.
}
\description{
This function creates a \code{gen_tibble} object for use in examples in
documentation.
}
