% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_gen_tibble.R
\name{example_gt}
\alias{example_gt}
\title{Create example gen_tibble}
\usage{
example_gt(
  type = c("gen_tbl", "grouped_gen_tbl", "grouped_gen_tbl_sf", "gen_tbl_sf")
)
}
\arguments{
\item{type}{a character string indicating the type of \code{gen_tibble} to create}
}
\value{
an example object of the class \code{gen_tbl}.
}
\description{
This function creates a \code{gen_tibble} object for use in examples in
documentation.
}
