% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fct_inorder_base}
\alias{fct_inorder_base}
\title{Reorder factor levels by first appearance}
\usage{
fct_inorder_base(f, ordered = FALSE)
}
\arguments{
\item{f}{A factor or character vector}

\item{ordered}{Logical, should the resulting factor be ordered?}
}
\value{
A factor with levels in the order they first appear
}
\description{
This function reorders the levels of a factor so that they appear in the order
they first appear in the data.  This is a drop-in replacement
for \code{forcats::fct_inorder()}
}
\examples{
x <- factor(c("b", "a", "c", "a", "b", "c", "b", "a"))
fct_inorder_base(x)
fct_inorder_base(x, ordered = TRUE)

}
\keyword{internal}
