% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_high_relatedness.R
\name{filter_high_relatedness}
\alias{filter_high_relatedness}
\title{Filter individuals based on a relationship threshold}
\usage{
filter_high_relatedness(matrix, .x = NULL, cutoff = NULL, verbose = FALSE)
}
\arguments{
\item{matrix}{a square symmetric matrix of individuals containing relationship coefficients}

\item{.x}{a \code{\link{gen_tibble}} object}

\item{cutoff}{a threshold over which}

\item{verbose}{boolean whether to report to screen}
}
\value{
a list where [\link{1}] is individual ID's to retain, [\link{2}]
is individual ID's to remove, and [\link{3}] is a boolean where individuals to keep
are TRUE and individuals to remove are FALSE
}
\description{
This function takes a matrix of x by y individuals containing relatedness
coefficients and returns the maximum set of individuals that contains no
relationships above the given threshold.
}
\details{
TODO this function needs a test
}
