% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flip.R
\name{flip}
\alias{flip}
\title{Flip the strand of alleles}
\usage{
flip(bases)
}
\arguments{
\item{bases}{a vector of characters. The only valid characters are "a", "t",
"c" and "g".}
}
\value{
a vector of recoded bases
}
\description{
This function takes a vector of bases and recodes them.
}
\details{
Note that it does not check that all bases are valid. Also, missing alleles
"m" are kept as "m".
}
\keyword{internal}
