% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_pop_fst.R
\name{format_fst_list}
\alias{format_fst_list}
\title{Format output for pairwise_fst_ functions based on type}
\usage{
format_fst_list(
  fst_list,
  .x,
  pairwise_combn,
  .group_levels,
  type,
  by_locus_type,
  by_locus,
  return_num_dem = FALSE
)
}
\arguments{
\item{fst_list}{list of fst values}

\item{type}{type of object to return One of "tidy" or "pairwise" for a
pairwise matrix of populations. Default is "tidy".
@param by_locus_type type of object to return. One of "tidy", "matrix" or
"list". Default is "tidy".
@param by_locus boolean, determining whether Fst should be returned by
loci(TRUE), or as a single genome wide value obtained by
taking the ratio
of the mean numerator and denominator (FALSE, the default).
@param group_combinations group combinations
@keywords internal}
}
\description{
Format output for pairwise_fst_ functions based on type
}
