% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loci_alt_freq.R
\name{format_grouped_output}
\alias{format_grouped_output}
\title{Function to format the matrix for grouped operations}
\usage{
format_grouped_output(out_mat, group_ids, loci_names, type)
}
\arguments{
\item{out_mat}{a matrix of the output}

\item{group_ids}{a vector of group ids}

\item{loci_names}{a vector of loci names}

\item{type}{a string indicating the type of output}
}
\value{
the data formatted as a tibble, list, or matrix
}
\description{
This function is used to format the matrix for grouped operations
}
\keyword{internal}
