% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotypes_count_na.R
\name{genotypes_count_na}
\alias{genotypes_count_na}
\title{Count the number of missing genotypes}
\usage{
genotypes_count_na(.x, alleles_as_unit = TRUE)
}
\arguments{
\item{.x}{a list of \code{\link[adegenet:SNPbin]{adegenet::SNPbin}} objects (usually the \code{genotype} column of
a \code{\link{gen_tibble}} object),}

\item{alleles_as_unit}{a logical indicating whether alleles are considered
as units (i.e., a diploid genotype equals two samples, a triploid, three,
etc.) or whether individuals are considered as units of information.}
}
\value{
a vector of counts of NAs
}
\description{
This function counts the number of missing genotypes for each SNP.
When ploidy varies across individuals, the outputs of this function depend
on whether the information units are individuals, or
alleles within individuals (see details).
}
