% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_packedancestry_code.R
\name{get_packedancestry_code}
\alias{get_packedancestry_code}
\title{Generate a packed ancestry code}
\usage{
get_packedancestry_code()
}
\value{
a matrix of raw values
}
\description{
This function returns a matrix of raw values (which can be used in a file
backed matrix) that represent the packed ancestry code. The columns
are the packed bytes from 0 to 255, and the rows are the genotypes for the
four individuals represented by the byte. Each two bits give the number of
alleles, with the combination 11 representing a missing value.
}
\keyword{internal}
