% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_admixture.R
\name{gt_admixture}
\alias{gt_admixture}
\title{Run ADMIXTURE from R}
\usage{
gt_admixture(
  x,
  k,
  crossval = FALSE,
  n_cores = 1,
  out = NULL,
  conda_env = "auto"
)
}
\arguments{
\item{x}{a \code{gen_tibble} or a character giving the path of the input PLINK bed file}

\item{k}{an integer giving the number of clusters}

\item{crossval}{boolean, should cross validation be used to assess the fit (defaults ot FALSE)}

\item{n_cores}{number of cores (defaults to 1)}

\item{out}{the output path for the results. It defaults to NULL, in which case, if \code{x} is a \code{gen_tibble}, the output
will be written to a temporary directory; esle if \code{x} is the path of a PLINK bed file,
the output will be saved in the same directory as the input file.}

\item{conda_env}{the name of the conda environment to use. If set to "auto", the default,
a copy from \code{tidygenclust} will be preferred if available, otherwise a local copy will
be used. "none" forces the use of a local copy, whilst any other string will
direct the function to use a custom conda environment.}
}
\value{
a list with the following elements:
\itemize{
\item \code{Q} a matrix with the admixture proportions
\item \code{loglik} the log likelihood of the model
\item \code{cv} the cross validation error (if \code{crossval} is TRUE)
}
}
\description{
This function runs ADMIXTURE, taking either a \code{gen_tibble} or a file as an input.
}
\details{
This is a wrapper for the command line program ADMIXTURE. It can either
use a binary present in the main environment, or use a copy installed in a conda
environment. If the package \code{tidygenclust} is installed, and its conda environment
has been set up with \code{ADMIXTURE} in it (the default), it will automatically
use that version unless you change \code{conda_env} to "none".
}
