% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_as_fbm.R
\name{gt_as_fbm}
\alias{gt_as_fbm}
\title{Convert a \code{gen_tibble} to a file backed matrix \code{FBM} from \code{bigstatsr}}
\usage{
gt_as_fbm(x)
}
\arguments{
\item{x}{a \code{\link{gen_tibble}}}
}
\value{
a \code{\link[bigstatsr:FBM]{FBM}} object from the \code{bigstatsr} package.
}
\description{
This function converts a \code{gen_tibble} object to a file-backed matrix (\code{FBM})
from the \code{bigstatsr} package. As \code{gen_tibble} objects use an \code{FBM}
internally, this simply returns the appropriate \code{FBM} object. Note that, for
tibble with imputed data, the genotype coding will depend on the impute
setting of the \code{gen_tibble} object (this can be checked with
\code{\link[=gt_uses_imputed]{gt_uses_imputed()}} and changed with \code{\link[=gt_set_imputed]{gt_set_imputed()}}).
}
\examples{
# Create a gen_tibble
example_gt <- load_example_gt("gen_tbl")
# Convert the gen_tibble to a file-backed matrix
example_fbm <- gt_as_fbm(example_gt)
# Check the class of the resulting object
class(example_fbm)
}
