% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_as_genlight.R
\name{gt_as_genlight}
\alias{gt_as_genlight}
\title{Convert a \code{gen_tibble} to a \code{genlight} object from \code{adegenet}}
\usage{
gt_as_genlight(x)
}
\arguments{
\item{x}{a \code{\link{gen_tibble}}, with population coded as 'population'}
}
\value{
a \code{genlight} object
}
\description{
This function converts a \code{gen_tibble} to a \code{genlight} object from \code{adegenet}
}
\examples{
example_gt <- load_example_gt("gen_tbl")

# Convert to genind
gt_genlight <- example_gt \%>\% gt_as_genlight()

# Check object class
class(gt_genlight)

}
