% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_to_aftable.R
\name{gt_extract_f2}
\alias{gt_extract_f2}
\title{Compute adn store blocked f2 statistics for a \code{gen_tibble}}
\usage{
gt_extract_f2(
  .x,
  outdir = NULL,
  cols_per_chunk = 10,
  blgsize = 0.05,
  quiet = FALSE
)
}
\arguments{
\item{.x}{the \code{gen_tibble}, appropriately filtered for individuals, populations and snps}

\item{outdir}{the directory where the f2 stats will be stored. If left NULL, a directory
names 'f2' will be created in the same path as the RDS fo the \code{gen_tibble}}

\item{cols_per_chunk}{Number of allele frequency chunks to store on disk.
Setting this to a positive integer makes the function slower,
but requires less memory. The default value for cols_per_chunk
in extract_afs is 10. Lower numbers will lower the memory requirement
but increase the time it takes.}

\item{blgsize}{SNP block size in Morgan. Default is 0.05 (5 cM). If \code{blgsize}
is 100 or greater, if will be interpreted as base pair distance rather than centimorgan distance.}

\item{quiet}{boolean on whether the progess should be silenced}
}
\value{
SNP metadata (invisibly)
}
\description{
This function prepares data for various \verb{ADMIXTOOLS 2} function, and it is
equivalent to \code{admixtools::extract_f2}. An important difference is that
the filtering for snps (e.g. by maf) or populations is not performed by this
function; it is expected that the \code{gen_tibble} has been filtered appropriately.
}
