% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_from_genlight.R
\name{gt_from_genlight}
\alias{gt_from_genlight}
\title{Convert a \code{genlight} object from adegenet to a \code{gen_tibble}}
\usage{
gt_from_genlight(x, backingfile = NULL, ...)
}
\arguments{
\item{x}{A \code{genlight} object}

\item{backingfile}{the path, including the file name without extension, for
backing files used to store the data (they will be given a .bk and .RDS
automatically). If \code{NULL} (default), backing files are placed in the
temporary directory.}

\item{...}{Additional arguments}
}
\value{
A \code{gen_tibble} object
}
\description{
This function converts a \code{genlight} object from the \code{adegenet} package to a
\code{gen_tibble} object
}
