% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_has_imputed.R
\name{gt_has_imputed}
\alias{gt_has_imputed}
\title{Checks if a \code{gen_tibble} has been imputed}
\usage{
gt_has_imputed(x)
}
\arguments{
\item{x}{a \code{gen_tibble}}
}
\value{
boolean TRUE or FALSE depending on whether the dataset has been imputed
}
\description{
This function checks if a dataset has been imputed. Note that having imputation
does not mean that the imputed values are used.
}
