% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_ibs.R
\name{gt_ibs}
\alias{gt_ibs}
\title{Compute the Identity by State Matrix for a \code{gen_tibble} object}
\usage{
gt_ibs(
  x,
  as_counts = FALSE,
  block_size = bigstatsr::block_size(length(show_loci_names(x))) * 2
)
}
\arguments{
\item{x}{a \code{gen_tibble} object.}

\item{as_counts}{whether the counts of similar alleles, rather than the proportion,
should be returned (FALSE by default).}

\item{block_size}{maximum number of loci read at once. More loci should improve speed,
but will tax memory.}
}
\value{
a list of two \link[bigstatsr:FBM-class]{bigstatsr::FBM} matrices, one of counts of IBS by alleles (i.e. 2\emph{n loci),
and one of valid alleles (i.e. 2}n_loci - 2*missing_loci)
}
\description{
This function computes the IBS matrix.
}
\details{
Note that monomorphic sites are currently counted. Should we filter
them beforehand? What does plink do?
}
