% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_load.R
\name{gt_load}
\alias{gt_load}
\title{Load a gen_tibble}
\usage{
gt_load(file_name = NULL)
}
\arguments{
\item{file_name}{the file name, including the full path. If it does not end
with '.gt', the extension will be added.}
}
\value{
a \code{\link{gen_tibble}}
}
\description{
Load a \code{gen_tibble} previously saved with \code{\link[=gt_save]{gt_save()}}.
}
