% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_order_loci.R
\name{gt_order_loci}
\alias{gt_order_loci}
\title{Order the loci table of a gen_tibble}
\usage{
gt_order_loci(.x, use_current_table = FALSE)
}
\arguments{
\item{.x}{a \link{'gen_tibble'}}

\item{use_current_table}{boolean, if FALSE (the default), the table will be
reordered; if TRUE, then the current loci table, which might have been reordered
manually, will be used, but only if the positions within each chromosome are
sequential}
}
\value{
a \link{'gen_tibble'}
}
\description{
This function reorders the loci table so that positions within a chromosome
are sequential. It also re-saves the genotypes into a new file backed matrix
with the new order, so that it can be used by functions such as \code{\link[=loci_ld_clump]{loci_ld_clump()}}
and \code{\link[=gt_pca_autoSVD]{gt_pca_autoSVD()}}. If the loci table is already ordered, the original
\code{gen_tibble} is returned.
}
