% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_save.R
\name{gt_save_light}
\alias{gt_save_light}
\title{a light version of gt_save that does not resave the bigSNP RDS or backing file, to be used internally
when creating a gen_tibble (if we have just created it, there is not need
to resave it)}
\usage{
gt_save_light(x, file_name = NULL, quiet = FALSE)
}
\arguments{
\item{x}{a \code{\link{gen_tibble}}}

\item{file_name}{the file name, including the full path. If it does not end
with \emph{.gt}, the extension will be added.}

\item{quiet}{boolean to suppress information about hte files}
}
\value{
the file name and path of the \emph{.gt} file, together with the \emph{.rds}
and \emph{.bk} files
}
\description{
a light version of gt_save that does not resave the bigSNP RDS or backing file, to be used internally
when creating a gen_tibble (if we have just created it, there is not need
to resave it)
}
\keyword{internal}
