% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_has_imputed.R
\name{gt_set_imputed}
\alias{gt_set_imputed}
\title{Sets a \code{gen_tibble} to use imputed data}
\usage{
gt_set_imputed(x, set = NULL)
}
\arguments{
\item{x}{a \code{gen_tibble}}

\item{set}{a boolean defining whether imputed data should be used}
}
\description{
This function sets or unsets the use of imputed data. For some analysis,
such as PCA, that does not allow for missing data, we have to use imputation,
but for other analysis it might be preferabble to allow for missing data.
}
