% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_extract_f2.R
\name{gt_to_aftable}
\alias{gt_to_aftable}
\title{Create an allele frequency table (aftable) for admixtools}
\usage{
gt_to_aftable(.x, adjust_pseudohaploid = TRUE, n_cores = bigstatsr::nb_cores())
}
\arguments{
\item{.x}{the \code{\link{gen_tibble}}}

\item{adjust_pseudohaploid}{Genotypes of pseudohaploid samples are
usually coded as \code{0} or \code{2}, even though only one allele is observed.
\code{adjust_pseudohaploid} ensures that the observed allele count increases
only by \code{1} for each pseudohaploid sample. If \code{TRUE} (default), samples
that don't have any genotypes coded as \code{1} among the first 1000 SNPs are
automatically identified as pseudohaploid. This leads to slightly more
accurate estimates of f-statistics. Setting this parameter to \code{FALSE}
treats all samples as diploid and is equivalent to the \emph{ADMIXTOOLS} \code{ inbreed: NO} option. Setting \code{adjust_pseudohaploid} to an integer \code{n}
will check the first \code{n} SNPs instead of the first 1000 SNPs.}

\item{n_cores}{Parallelize computation across \code{n_cores} cores.}
}
\value{
a list of 3 elements: afs, counts, and snp
}
\description{
This function is equivalent to \code{anygeno_to_aftable()}, but the aftable is
created directly from the \code{gen_tibble}.
}
\keyword{internal}
