% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_update_backingfile.R
\name{gt_update_backing_file}
\alias{gt_update_backing_file}
\title{Update the backing matrix}
\usage{
gt_update_backing_file(
  .x,
  backingfile = NULL,
  chunk_size = NULL,
  quiet = FALSE
)
}
\arguments{
\item{.x}{a \code{gen_tibble} object}

\item{backingfile}{@param backingfile the path, including the file name without extension,
for backing files used to store the data (they will be given a .bk
and .RDS automatically). If left to NULL (the default), the file name
will be based on the name f the current backing file.}

\item{quiet}{boolean to suppress information about the files}
}
\value{
a \code{\link{gen_tibble}} with a backing file (i.e. a new File Backed Matrix)
}
\description{
This functions forces a re-write of the file backing matrix to match the
\code{\link{gen_tibble}}. Individuals and loci are subsetted and reordered according to
the current state of the \code{gen_tibble}.
}
