% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_write_lea_geno.R
\name{gt_write_lea_geno}
\alias{gt_write_lea_geno}
\title{Write a geno file for sNMF from the LEA package}
\usage{
gt_write_lea_geno(x, file = NULL)
}
\arguments{
\item{x}{a \code{\link{gen_tibble}}}

\item{file}{the .geno filename with a path, or NULL (the default) to use the
location of the backing files.}
}
\value{
the path of the .geno file
}
\description{
This function writes a geno file fom a \code{\link{gen_tibble}}. Unless a file path is given,
a file with suffix .geno is written in the same location as the .rds and .bk
files that underpin the \code{\link{gen_tibble}}.
}
