% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_pairwise.R
\name{heatmap_pairwise}
\alias{heatmap_pairwise}
\title{Pairwise distance heatmap}
\usage{
heatmap_pairwise(x, order = NULL)
}
\arguments{
\item{x}{A square numeric matrix OR a data.frame/tibble with 3 columns:
names1, names2, value (half-matrix representation).}

\item{order}{Optional ordering specification. Can be:
\itemize{
\item \code{NULL} (default): preserve the existing order.
\item A character vector containing all row/column names in the desired
order.
\item A function taking the matrix \code{x} as input and returning either:
\itemize{
\item an ordering vector of indices, or
\item an object with an \code{order} component (e.g. from
\code{\link[stats:hclust]{stats::hclust()}}).
}
}}
}
\value{
A ggplot object.
}
\description{
Plot a square heatmap from either a full symmetric distance matrix or a tidy
tibble of distances. Rows and columns can optionally be reordered using
either an explicit ordering vector or a function that computes an ordering
(e.g. hierarchical clustering). Dendrograms can be added to the heatmap with
\code{\link[=heatmap_add_dendro]{heatmap_add_dendro()}}. Note that any decoration (e.g. axis labels, fill
scale) should be added to the heatmap before adding dendrograms, as the
latter will modify the plot structure and make it difficult to add further
customization afterwards.
}
\examples{
example_gt <- load_example_gt("gen_tbl")

# With a pairwise matrix:
fst_dist <- example_gt \%>\%
  group_by(population) \%>\%
  pairwise_pop_fst(method = "Nei87", type = "pairwise")

heatmap_pairwise(fst_dist)

# Explicit ordering
heatmap_pairwise(
  fst_dist,
  order = rev(rownames(fst_dist))
) +
  ggplot2::scale_fill_viridis_c()

# Hierarchical clustering
heatmap_pairwise(
  fst_dist,
  order = function(x) {
    stats::hclust(stats::as.dist(x), method = "complete")
  }
)

# With a tidy tibble of distances:
fst_tidy <- example_gt \%>\%
  group_by(population) \%>\%
  pairwise_pop_fst(method = "Nei87", type = "tidy")

heatmap_pairwise(fst_tidy)

heatmap_pairwise(
  fst_tidy,
  order = function(x) {
    stats::hclust(stats::as.dist(x))
  }
)

}
\seealso{
\code{\link[=heatmap_add_dendro]{heatmap_add_dendro()}}
}
