% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierfstat_basic_stats.R
\name{hierfstat_basic_stats}
\alias{hierfstat_basic_stats}
\title{Compute basic population statistics}
\usage{
hierfstat_basic_stats(.x, n_cores = bigstatsr::nb_cores())
}
\arguments{
\item{.x}{a \code{\link{gen_tibble}} (usually grouped, as obtained by using \code{\link[dplyr:group_by]{dplyr::group_by()}}; use on
a single population will return a number of quantities as NA/NaN)}

\item{n_cores}{number of cores to be used, it defaults to \code{\link[bigstatsr:reexports]{bigstatsr::nb_cores()}}}
}
\value{
a tibble of population statistics, with populations as rows and statistics as columns
}
\description{
This function computes basic population statistics, following the formulas used in \code{hietfstat}.
}
