% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_pseudohaploid.R
\name{identify_pseudohaploids}
\alias{identify_pseudohaploids}
\title{Identify pseudohaploids}
\usage{
identify_pseudohaploids(x, n_test = 1000)
}
\arguments{
\item{x}{the gen_tibble}

\item{n_test}{the number of loci being tested}
}
\value{
a numeric vector of ploidy
}
\description{
Pseudohaploids are coded as all homozygotes; we find them by checking the
first 'n_test' loci and call a pseudohaploid if they have zero heterozygosity
(this is the same strategy employed in admixtools)
}
\keyword{internal}
