% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indiv_ploidy.R
\name{indiv_ploidy}
\alias{indiv_ploidy}
\alias{indiv_ploidy.tbl_df}
\alias{indiv_ploidy.vctrs_bigSNP}
\alias{indiv_ploidy.grouped_df}
\title{Estimate individual ploidy}
\usage{
indiv_ploidy(.x, ...)

\method{indiv_ploidy}{tbl_df}(.x, ...)

\method{indiv_ploidy}{vctrs_bigSNP}(.x, ...)

\method{indiv_ploidy}{grouped_df}(.x, ...)
}
\arguments{
\item{.x}{a vector of class \code{vctrs_bigSNP} (usually the \code{genotype} column of
a \code{\link{gen_tibble}} object),
or a \code{\link{gen_tibble}}.}

\item{...}{currently unused.}
}
\value{
a vector of ploidy, one per individuals in the \code{\link{gen_tibble}}
}
\description{
Estimate missingnes for each individual (i.e. the frequency of
missing genotypes in an individual).
}
