% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_example_gt.R
\name{load_example_gt}
\alias{load_example_gt}
\title{Load example gen_tibble}
\usage{
load_example_gt(
  type = c("gen_tbl", "grouped_gen_tbl", "grouped_gen_tbl_sf", "gen_tbl_sf")
)
}
\arguments{
\item{type}{a character string indicating the type of \code{gen_tibble} to create:
\itemize{
\item "gen_tbl": a basic gen_tibble with genotype data and metadata
\item "grouped_gen_tbl": same as "gen_tbl" but grouped by population
\item "grouped_gen_tbl_sf": adds spatial features (longitude/latitude)
and groups by population
\item "gen_tbl_sf": adds spatial features without grouping
}}
}
\value{
an example object of the class \code{gen_tbl}.
}
\description{
This function creates a \code{gen_tibble} object for use in examples in
documentation.
}
\examples{
# This function creates an example gen_tibble object
example_gt <- load_example_gt("gen_tbl")
}
