% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loci_chromosomes.R
\name{loci_chromosomes}
\alias{loci_chromosomes}
\alias{loci_chromosomes.tbl_df}
\alias{loci_chromosomes.vctrs_bigSNP}
\title{Get the chromosomes of loci in a \code{gen_tibble}}
\usage{
loci_chromosomes(.x, ...)

\method{loci_chromosomes}{tbl_df}(.x, ...)

\method{loci_chromosomes}{vctrs_bigSNP}(.x, ...)
}
\arguments{
\item{.x}{a \code{\link{gen_tibble}}, or a vector of class \code{vctrs_bigSNP} (usually the
\code{genotype} column of a \code{\link{gen_tibble}} object).}

\item{...}{currently unused.}
}
\value{
a character vector of chromosomes
}
\description{
Extract the loci chromosomes from a \code{gen_tibble} (or directly from its
\code{genotype} column).
}
\examples{
example_gt <- example_gt()
example_gt \%>\% loci_chromosomes()
}
