% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loci_freq.R
\name{loci_freq}
\alias{loci_freq}
\alias{loci_freq.tbl_df}
\alias{loci_freq.list}
\alias{loci_freq.grouped_df}
\title{Estimate allele frequencies at each each locus}
\usage{
loci_freq(.x, ...)

\method{loci_freq}{tbl_df}(.x, ..., minor = TRUE, alleles_as_units = TRUE, use_c = FALSE)

\method{loci_freq}{list}(.x, ..., minor = TRUE, alleles_as_units = TRUE, use_c = FALSE)

\method{loci_freq}{grouped_df}(.x, ..., minor = TRUE, alleles_as_units = TRUE, use_c = FALSE)
}
\arguments{
\item{.x}{a list of \code{\link[adegenet:SNPbin]{adegenet::SNPbin}} objects (usually the \code{genotype} column of
a \code{\link{gen_tibble}} object),
or a \code{\link{gen_tibble}}.}

\item{...}{other arguments passed to specific methods.}

\item{minor}{a logical indicating whether we should give the frequencies of
the minor allele (TRUE, the default). If FALSE, the frequencies of the
alternate allele are given.}

\item{alleles_as_units}{a logical indicating whether alleles are considered
as units (i.e., a diploid genotype equals two samples, a triploid, three,
etc.) or whether individuals are considered as units of information.}

\item{use_c}{a logical indicating whether compiled C code should be used
(TRUE) or not (FALSE, default).}
}
\value{
a vector of frequencies, one per locus
}
\description{
Estimate the frequency of the alternate allele at each locus.
}
