% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loci_names.R
\name{loci_names}
\alias{loci_names}
\alias{loci_names.tbl_df}
\alias{loci_names.vctrs_bigSNP}
\title{Get the names of loci in a \code{gen_tibble}}
\usage{
loci_names(.x, ...)

\method{loci_names}{tbl_df}(.x, ...)

\method{loci_names}{vctrs_bigSNP}(.x, ...)
}
\arguments{
\item{.x}{a vector of class \code{vctrs_bigSNP} (usually the \code{genotype} column of
a \code{\link{gen_tibble}} object),
or a \code{\link{gen_tibble}}.}

\item{...}{currently unused.}
}
\value{
a character vector of names
}
\description{
Extract the loci names from a  \code{gen_tibble} (or directly from its \code{genotype}
column).
}
\examples{
example_gt <- example_gen_tibble()
example_gt \%>\% loci_names()
}
