% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loci_qc_report.R
\name{loci_qc_report}
\alias{loci_qc_report}
\title{Create a Quality Control report for loci}
\usage{
loci_qc_report(.x, ...)
}
\arguments{
\item{.x}{a \code{\link{gen_tibble}} object.}

\item{...}{further arguments to pass to \code{\link[HardyWeinberg:HWExact]{HardyWeinberg::HWExact()}} for changing
the HWE test.}
}
\value{
a tibble with 3 elements: maf, missingness and hwe_p
}
\description{
Return QC information to assess loci (MAF, missingness and HWE test).
}
