% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loci_transversions.R
\name{loci_transversions}
\alias{loci_transversions}
\alias{loci_transversions.tbl_df}
\alias{loci_transversions.vctrs_bigSNP}
\alias{loci_transversions.grouped_df}
\title{Estimates the sum of genotypes at each each locus}
\usage{
loci_transversions(.x, ...)

\method{loci_transversions}{tbl_df}(.x, ...)

\method{loci_transversions}{vctrs_bigSNP}(.x, ...)

\method{loci_transversions}{grouped_df}(.x, ...)
}
\arguments{
\item{.x}{a vector of class \code{vctrs_bigSNP} (usually the \code{genotype} column of
a \code{\link{gen_tibble}} object),
or a \code{\link{gen_tibble}}.}

\item{...}{other arguments passed to specific methods.}
}
\value{
a vector of frequencies, one per locus
}
\description{
Estimate the sum of the alternate allele at each locus. This is unlikely to be useful
directly, but it is used by other functions that compute various statistics.
}
