% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyverse.R
\name{mutate.gen_tbl}
\alias{mutate.gen_tbl}
\title{A mutate method for \code{gen_tibble} objects}
\usage{
\method{mutate}{gen_tbl}(..., deparse.level = 1)
}
\arguments{
\item{...}{a gen_tibble and a data.frame or tibble}

\item{deparse.level}{an integer controlling the construction of column names.}
}
\value{
a \code{gen_tibble}
}
\description{
A mutate method for \code{gen_tibble} objects
}
\examples{
example_gt <- load_example_gt("gen_tbl")

# Add a new column
example_gt \%>\% mutate(region = "East")
}
