% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyverse.R
\name{mutate.grouped_gen_tbl}
\alias{mutate.grouped_gen_tbl}
\title{A mutate method for grouped \code{gen_tibble} objects}
\usage{
\method{mutate}{grouped_gen_tbl}(..., deparse.level = 1)
}
\arguments{
\item{...}{a gen_tibble and a data.frame or tibble}

\item{deparse.level}{an integer controlling the construction of column names.}
}
\value{
a grouped \code{gen_tibble}
}
\description{
A mutate method for grouped \code{gen_tibble} objects
}
\examples{
test_gt <- load_example_gt("grouped_gen_tbl")
test_gt \%>\% mutate(region = "East")
test_gt <- load_example_gt("grouped_gen_tbl_sf")
test_gt \%>\% mutate(region = "East")
}
