% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_pop_pbs.R
\name{pairwise_pop_pbs}
\alias{pairwise_pop_pbs}
\title{Compute the Population Branch Statistics for each combination of populations}
\usage{
pairwise_pop_pbs(
  .x,
  type = c("matrix"),
  fst_method = c("Hudson"),
  return_fst = FALSE
)
}
\arguments{
\item{.x}{A grouped \code{gen_tibble}}

\item{type}{type of object to return. Currently defaults to "matrix".}

\item{fst_method}{A character string specifying the method to use for
computing Fst. Currently only "Hudson" is available.}

\item{return_fst}{A logical value indicating whether to return the Fst values
along with the PBS values. Default is \code{FALSE}.}
}
\value{
A tibble with the following columns:
\itemize{
\item \code{pbs_a.b.c}: the PBS value for population a given b & c (there
will be multiple such columns covering all 3 way combinations of
populations in the grouped \code{gen_tibble} object)
\item \code{fst_a.b}: the Fst value for population a and b, if \code{return_fst} is TRUE
}
}
\description{
The function computes the population branch statistics (PBS) for each
combination of populations at each locus. The PBS is a measure of the genetic
differentiation between one focal population and two reference populations,
and is used to identify outlier loci that may be under selection.
}
\examples{
example_gt <- example_gt()

# We can compute the PBS for all populations using "Hudson" method
example_gt \%>\%
  group_by(population) \%>\%
  pairwise_pop_pbs(fst_method = "Hudson")
}
\references{
Yi X, et al. (2010) Sequencing of 50 human exomes reveals
adaptation to high altitude. Science 329: 75-78.
}
