% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_packedancestry_code.R
\name{parse_2bit_groups}
\alias{parse_2bit_groups}
\title{Parse a byte into 2-bit groups}
\usage{
parse_2bit_groups(byte)
}
\arguments{
\item{byte}{A raw value representing a byte (0-255).}
}
\value{
A vector of integers representing the 4 groups of 2 bits.
}
\description{
This function takes a byte (raw value) and extracts 4 groups of 2 bits each.
Following the convention by \emph{packedancestry}, we start from the left (i.e the
higher value bits) and move to the right (i.e the lower value bits).
}
\keyword{internal}
