% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_pop_pbs.R
\name{pbs_one_triplet}
\alias{pbs_one_triplet}
\title{Internal function to compute pbs for a single triplet of populations}
\usage{
pbs_one_triplet(pops, fst_values)
}
\arguments{
\item{pops}{A vector of 3 population names}

\item{fst_values}{A matrix of Fst values (rows are markers/windows, columns
are named as fst_pop1.pop2, fst_pop1.pop3, fst_pop2.pop3)}
}
\value{
A tibble with the PBS values for the triplet of populations
}
\description{
Internal function to compute pbs for a single triplet of populations
}
\keyword{internal}
