% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_pairwise_fst.R
\name{pop_pairwise_fst}
\alias{pop_pairwise_fst}
\title{Compute pairwise Hudson Fst}
\usage{
pop_pairwise_fst(.x, by_locus = FALSE)
}
\arguments{
\item{.x}{a grouped \code{\link{gen_tibble}} (as obtained by using \code{\link[dplyr:group_by]{dplyr::group_by()}})}

\item{by_locus}{boolean, determining whether Fst should be returned by locus(TRUE),
or as a single genome wide value obtained by taking the ratio of the mean numerator
and denominator (FALSE, the default).}
}
\value{
a tibble with each pairwise combination as a row
}
\description{
This function computes pairwise Fst using the formulation by Hudson.
}
