% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q_matrix.R
\name{q_matrix}
\alias{q_matrix}
\title{Read and structure .Q files or existing matrices as 'q_matrix' objects.}
\usage{
q_matrix(x)
}
\arguments{
\item{x}{a filepath or matrix or tibble}
}
\value{
a list of q_matrices
}
\description{
Read and structure .Q files or existing matrices as 'q_matrix' objects.
}
