% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loci_qc_report.R
\name{qc_report_loci}
\alias{qc_report_loci}
\title{Create a Quality Control report for loci}
\usage{
qc_report_loci(.x, ...)
}
\arguments{
\item{.x}{a \code{\link{gen_tibble}} object.}

\item{...}{currently unused
the HWE test.}
}
\value{
a tibble with 3 elements: maf, missingness and hwe_p
}
\description{
Return QC information to assess loci (MAF, missingness and HWE test).
}
