% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_rbind_report.R
\name{summary.rbind_report}
\alias{summary.rbind_report}
\title{Print a summary of a merge report}
\usage{
summary.rbind_report(report, ref_label = "reference", target_label = "target")
}
\arguments{
\item{report}{a list generated by \code{\link[=rbind_dry_run]{rbind_dry_run()}}}

\item{ref_label}{the label for the reference dataset (defaults to "reference")}

\item{target_label}{the label for the target dataset (defaults to "target")}
}
\value{
NULL (prints a summary to the console)
}
\description{
This function creates a summary of the merge report generated by
\code{\link[=rbind_dry_run]{rbind_dry_run()}}
}
