% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbind_gen_tibble.R
\name{rbind_gen_tbl}
\alias{rbind_gen_tbl}
\title{Combine two gen_tibbles}
\usage{
rbind_gen_tbl(
  ref,
  target,
  as_is = FALSE,
  flip_strand = FALSE,
  remove_ambiguous = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{ref}{either a \code{\link{gen_tibble}} object, or the path to the PLINK file;
the alleles in this objects will
be used as template to flip the ones in \code{target} and/or
swap their order as necessary.}

\item{target}{either a \code{\link{gen_tibble}} object, or the path to the PLINK file
(saved in raw format, see details in \code{\link[=read_plink_raw]{read_plink_raw()}}.}

\item{as_is}{boolean determining whether the loci should be left as they are
before merging. If FALSE (the defaults), \code{rbind} will attempt to subset and
swap alleles as needed.}

\item{flip_strand}{boolean on whether strand flipping should be checked to
match the two datasets. It defaults to FALSE}

\item{remove_ambiguous}{boolean whether ambiguous SNPs (i.e. a/t and c/g)
should be removed. It defaults to FALSE}

\item{quiet}{boolean whether to omit reporting to screen}
}
\value{
a \code{\link{gen_tibble}} with the merged data.
}
\description{
This function combined two \link{gen_tibble}s. By defaults, it subsets the loci
and swaps ref and alt alleles to make the two datasets compatible (this
behaviour can be switched off with \code{as_is}).
The first object is used as a "reference2 , and SNPs
in the other dataset will be flipped and/or alleles swapped
as needed. SNPs that have different alleles in the two datasets
(i.e. triallelic) will also be
dropped. There are also options (NOT default) to attempt strand flipping to
match alleles (often needed in human datasets from different SNP chips),
and remove ambiguous alleles (c/g and a/t) where the correct strand can not
be guessed.
}
