% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_plink_raw.R
\name{read_plink_raw}
\alias{read_plink_raw}
\title{Read data in PLINK raw format and store it to a \code{gen_tibble}}
\usage{
read_plink_raw(
  file,
  map_file = NULL,
  quiet = FALSE,
  chunk_size = 1000,
  parallel = FALSE,
  n_cores = NULL,
  ...
)
}
\arguments{
\item{file}{a character string giving the path to the file to convert,
with the extension ".raw"}

\item{map_file}{character string indicating the path to a ".map" file,
which contains information about the SNPs (chromosome, position).}

\item{quiet}{logical stating whether a conversion messages should be printed
(TRUE) or not (FALSE, default).}

\item{chunk_size}{an integer indicating the number of genomes to be read at a time;
larger values require more RAM but decrease the time needed to read the data.}

\item{parallel}{a logical indicating whether multiple cores -if available-
should be used for the computations (TRUE, default), or not (FALSE);
requires the package parallel to be installed (see details).}

\item{n_cores}{if parallel is TRUE, the number of cores to be used in the
computations; if NULL, then the maximum number of cores
available on the computer is used.}

\item{...}{other arguments to be passed to other functions - currently not used.}
}
\value{
an object of the class \code{gen_tbl}.
}
\description{
This function reads a PLINK raw file (and optionally a map file) and stores
the data as a \code{gen_tibble}. The raw file should be generated with the flags
\code{--include-alt} and \code{--recodeA}. The map file is generated with \code{--recode} (in other words,
you will need to run PLINK twice to generate the necessary files).
}
\examples{
raw_path <- system.file("extdata/san.raw", package = "tidypopgen")
map_path <- system.file("extdata/san.map", package = "tidypopgen")
san_gen_tbl <- read_plink_raw(file = raw_path, map_file = map_path)

}
