% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_plink_raw_loci.R
\name{read_plink_raw_loci}
\alias{read_plink_raw_loci}
\title{Read loci information directly from PLINK raw files}
\usage{
read_plink_raw_loci(file)
}
\arguments{
\item{file}{the path to the PLINK raw file}
}
\value{
an tibble with the loci inofrmation
}
\description{
This function extract locus names and alleles from a PLINK raw file, without
the need to build a full \code{gen_tibble} object. This function returns a tibble
equivalent to a subset of \code{loci} attribute of the \code{genotypes} column (i.e
missing 'chromosome' and 'position')
}
\keyword{internal}
