% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_q_matrix.R
\name{read_q_matrix_list}
\alias{read_q_matrix_list}
\title{Tidy ADMXITURE output files into plots}
\usage{
read_q_matrix_list(x, data)
}
\arguments{
\item{x}{the name of a directory containing .Q files}

\item{data}{An associated tibble (e.g. a \code{\link{gen_tibble}}), with the individuals in the same order as the data used to
generate the Q matrix}
}
\value{
a list of \code{q_matrix} objects to plot
}
\description{
Takes the name of a directory containing .Q file outputs, and
produces a list of tidied tibbles ready to plot.
}
