% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_int_names.R
\name{row.names<-.matrix_int_names}
\alias{row.names<-.matrix_int_names}
\title{Set row names for matrix_int_names}
\usage{
\method{row.names}{matrix_int_names}(x) <- value
}
\arguments{
\item{x}{A matrix_int_names object}

\item{value}{Integer or character vector of row names, or NULL}
}
\value{
The modified matrix_int_names object
}
\description{
Set row names for matrix_int_names
}
