% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_int_names.R
\name{row_names}
\alias{row_names}
\alias{row_names<-}
\alias{row_names<-.default}
\alias{row_names<-.matrix_int_names}
\title{Get or set row names for matrix_int_names}
\usage{
row_names(x)

row_names(x) <- value

\method{row_names}{default}(x) <- value

\method{row_names}{matrix_int_names}(x) <- value
}
\arguments{
\item{x}{A matrix_int_names object}

\item{value}{Integer or character vector of row names, or NULL}
}
\value{
Integer or character vector of row names, or NULL

The modified matrix_int_names object
}
\description{
Returns integer names if present, otherwise character dimnames. Use instead
of \code{rownames()} for guaranteed dispatch.
}
\seealso{
Other matrix_int_names_functions:
\code{\link[=[.matrix_int_names]{[.matrix_int_names()}},
\code{\link[=as.data.frame.matrix_int_names]{as.data.frame.matrix_int_names()}},
\code{\link[=col_names]{col_names()}},
\code{\link[=matrix_int_names]{matrix_int_names()}},
\code{\link[=print.matrix_int_names]{print.matrix_int_names()}}
}
\concept{matrix_int_names_functions}
