% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_loci.R
\name{select_loci}
\alias{select_loci}
\title{The \code{select_if} verb for \code{loci}}
\usage{
select_loci(.data, .sel_arg, .swap_arg = NULL, .swap_if_arg = NULL)
}
\arguments{
\item{.data}{a \code{gen_tibble}}

\item{.sel_arg}{one unquoted expression}

\item{.swap_arg}{not used yet}

\item{.swap_if_arg}{not used yet}
}
\value{
a list of \code{SNPbin} object that have been subsetted. Variable names
can be used as if they were positions in the data frame, so expressions
like x:y can be used to select a range of variables.
}
\description{
An equivalent to \code{\link[dplyr:select_all]{dplyr::select_if()}} that works on the \code{genotype} column
of a \code{gen_tibble}. Tidy evaluation should work as expected.
}
