% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_swap_snpbin.R
\name{select_swap_snpbin}
\alias{select_swap_snpbin}
\title{Select and swap SNPs in a SNP bin}
\usage{
select_swap_snpbin(x, sel_indices, to_swap)
}
\arguments{
\item{x}{a \link[adegenet:SNPbin]{adegenet::SNPbin} object.}

\item{sel_indices}{the indices and order in which loci should be rearranged,
or a boolean of the appropriate length}

\item{to_swap}{indices or a boolean vector of SNPs to swap}
}
\value{
an edited \link[adegenet:SNPbin]{adegenet::SNPbin} object.
}
\description{
This function first swaps the major and minor alleles for loci set at TRUE
in to_swap, and then keeps
}
\details{
Note that this function does not check that \code{to_swap} has the correct
length, nor that \code{sel_indices} contains valid indices (or it is a boolean
of the appropriate length). This is to optmise speed.
}
\keyword{internal}
