% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_swap_snpbin_list.R
\name{select_swap_snpbin_list}
\alias{select_swap_snpbin_list}
\title{Filter, flip and swap SNPs in a genlight object}
\usage{
select_swap_snpbin_list(x, new_order, to_flip, to_swap, missing_allele)
}
\arguments{
\item{x}{a \link[adegenet:SNPbin]{adegenet::SNPbin} object.}

\item{new_order}{a numeric vector with the new order of the SNPs (NAs will
be excluded)}

\item{to_flip}{a boolean vector of SNPs to flip}

\item{to_swap}{a boolean vector of SNPs to swap}

\item{missing_allele}{a character vector of alternate alleles to fill in
missing alleles (NA for loci without missing alleles)}
}
\value{
an edited \link[adegenet:SNPbin]{adegenet::SNPbin} object.
}
\description{
This function takes a genlight object and edits its SNPs
}
\details{
Note that this function does not check that to_keep and to_swap have the correct
length
}
\keyword{internal}
