% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_loci.R
\name{show_loci}
\alias{show_loci}
\alias{show_loci.gen_tbl}
\title{Show the loci information of a \code{gen_tibble}}
\usage{
show_loci(x, ...)

\method{show_loci}{gen_tbl}(x, ...)
}
\arguments{
\item{x}{a \code{\link{gen_tibble}} object.}

\item{...}{currently unused}
}
\value{
a \code{\link[tibble:tibble]{tibble::tibble}} of information (see\code{\link{gen_tibble}} for details
on compulsory columns that will always be present)
}
\description{
Extract the information on loci from a  \code{gen_tibble}.
}
