% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_loci_names.R
\name{show_loci_names}
\alias{show_loci_names}
\alias{show_loci_names.tbl_df}
\alias{show_loci_names.list}
\title{Show the loci names of a \code{gen_tibble}}
\usage{
show_loci_names(.x, ...)

\method{show_loci_names}{tbl_df}(.x, ...)

\method{show_loci_names}{list}(.x, ...)
}
\arguments{
\item{.x}{a list of \code{\link[adegenet:SNPbin]{adegenet::SNPbin}} objects (usually the \code{genotype} column of
a \code{\link{gen_tibble}} object),
or a \code{\link{gen_tibble}}.}

\item{...}{currently unused.}
}
\value{
a character vector of names
}
\description{
Extract the loci names from a  \code{gen_tibble} (or directly from its \code{genotype}
column).
}
