% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_ploidy.R
\name{show_ploidy}
\alias{show_ploidy}
\alias{show_ploidy.gen_tbl}
\alias{show_ploidy.list}
\title{Show the loci information of a \code{gen_tibble}}
\usage{
show_ploidy(.x, ...)

\method{show_ploidy}{gen_tbl}(.x, .col, ...)

\method{show_ploidy}{list}(.x, ...)
}
\arguments{
\item{.x}{a list of \code{\link[adegenet:SNPbin]{adegenet::SNPbin}} objects (usually the \code{genotype} column of
a \code{\link{gen_tibble}} object),
or a \code{\link{gen_tibble}}.}

\item{...}{currently unused.}

\item{.col}{if \code{.x} is a \code{\link{gen_tibble}}, the column containing the genotypes
(usually \code{genotypes})}
}
\value{
a \code{\link[tibble:tibble]{tibble::tibble}} of information (see\code{\link{gen_tibble}} for details
on compulsory columns that will always be present)
}
\description{
Extract the information on ploidy from a \code{gen_tibble}.
}
