% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windos_stats_generic.R
\name{slider_operator}
\alias{slider_operator}
\title{Internal function to choose the slider operator for windowing}
\usage{
slider_operator(operator, size_unit)
}
\arguments{
\item{operator}{The operator to use for the window statistics.}

\item{size_unit}{The unit of the window size. Either "snp" or "bp".}
}
\value{
The function to use for the window statistics.
}
\description{
Internal function to choose the slider operator for windowing
}
\keyword{internal}
