% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp_ibs_r.R
\name{snp_ibs_r}
\alias{snp_ibs_r}
\title{Compute the Identity by State Matrix for a bigSNP object}
\usage{
snp_ibs_r(
  X,
  row.ind = bigstatsr::rows_along(X),
  col.ind = bigstatsr::cols_along(X),
  as.counts = FALSE,
  block.size = bigstatsr::block_size(nrow(X))
)
}
\arguments{
\item{X}{a \link[bigstatsr:FBM.code256-class]{bigstatsr::FBM.code256} matrix (as found in the \code{genotypes}
slot of a \link[bigsnpr:bigSNP-class]{bigsnpr::bigSNP} object).}

\item{row.ind}{An optional vector of the row indices that are used.
If not specified, all rows are used. Don't use negative indices.}

\item{col.ind}{An optional vector of the column indices that are used. If not
specified, all columns are used. Don't use negative indices.}

\item{as.counts}{whether the counts of similar alleles, rather than the proportion,
should be returned (FALSE by default).}

\item{block.size}{maximum number of columns read at once.}
}
\description{
This function computes the IBS matrix. Missing genotypes will have to be
imputed before using this function. Computing the IBS of 1k individuals for
a million markers will take over 10 mins, so the function will not scale to
huge datasets. But for smaller dataset, it allows estimating the IBS directly
in R.
}
