% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp_king.R
\name{snp_king}
\alias{snp_king}
\title{Compute the KING-robust Matrix for a bigSNP object}
\usage{
snp_king(
  X,
  ind.row = bigstatsr::rows_along(X),
  ind.col = bigstatsr::cols_along(X),
  block.size = bigstatsr::block_size(nrow(X)) * 4
)
}
\arguments{
\item{X}{a \link[bigstatsr:FBM.code256]{bigstatsr::FBM.code256} matrix (as found in the \code{genotypes}
slot of a \link[bigsnpr:bigSNP]{bigsnpr::bigSNP} object).}

\item{ind.row}{An optional vector of the row indices that are used.
If not specified, all rows are used. Don't use negative indices.}

\item{ind.col}{An optional vector of the column indices that are used. If not
specified, all columns are used. Don't use negative indices.}

\item{block.size}{maximum number of columns read at once.}
}
\value{
a square symmetrical matrix of relationship coefficients between
individuals
}
\description{
This function computes the KING-robust estimator of kinship, reimplementing
the KING algorithm of Manichaikul et al. (2010).
}
\examples{
example_gt <- load_example_gt("gen_tbl")

X <- attr(example_gt$genotypes, "fbm")
snp_king(X)

# Compute for individuals 1 to 5
snp_king(X, ind.row = 1:5, ind.col = 1:5)

# Adjust block size
snp_king(X, block.size = 2)

}
\references{
Manichaikul, A. et al. (2010) Robust relationship inference in
genome-wide association studies. Bioinformatics, 26(22), 2867–2873.
https://doi.org/10.1093/bioinformatics/btq559.
}
