% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp_king_r.R
\name{snp_king_r}
\alias{snp_king_r}
\title{Compute the KING-robust Matrix for a bigSNP object}
\usage{
snp_king_r(
  X,
  row.ind = bigstatsr::rows_along(X),
  col.ind = bigstatsr::cols_along(X),
  block.size = bigstatsr::block_size(nrow(X))
)
}
\arguments{
\item{X}{a \link[bigstatsr:FBM.code256-class]{bigstatsr::FBM.code256} matrix (as found in the \code{genotypes}
slot of a \link[bigsnpr:bigSNP-class]{bigsnpr::bigSNP} object).}

\item{row.ind}{An optional vector of the row indices that are used.
If not specified, all rows are used. Don't use negative indices.}

\item{col.ind}{An optional vector of the column indices that are used. If not
specified, all columns are used. Don't use negative indices.}

\item{block.size}{maximum number of columns read at once.}
}
\description{
This function computes the KING-robust estimator of kinship. Missing genotypes will have to be
imputed before using this function. Computing the IBS of 1k individuals for
a million markers will take over 10 mins, so the function will not scale to
huge datasets. But for smaller dataset, it allows estimating the IBS directly
in R.
}
