% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snpbin_list_n.R
\name{snpbin_list_n}
\alias{snpbin_list_n}
\title{Count the number of alleles or individuals per locus}
\usage{
snpbin_list_n(.x, alleles_as_units = TRUE)
}
\arguments{
\item{.x}{a list of \code{\link[adegenet:SNPbin]{adegenet::SNPbin}} objects (usually the \code{genotype} column of
a \code{\link{gen_tibble}} object),}

\item{alleles_as_units}{a logical indicating whether alleles are considered
as units (i.e., a diploid genotype equals two samples, a triploid, three,
etc.) or whether individuals are considered as units of information.}
}
\value{
a vector of counts of alleles or individuals
}
\description{
This function counts the number of alleles or individual for each locus
(excluding NAs for any given genotype).
It is not
meant to be used as a general summary, but it is used within other functions to
compute intermediate quantities.
When ploidy varies across individuals, the outputs of this function depend
on whether the information units are individuals, or
alleles within individuals (see details for \code{\link[=snpbin_list_sums]{snpbin_list_sums()}}).
}
\details{
This function is a modified version
of \link[adegenet:glAux]{adegenet::glNA}, recoded to work on lists of \code{SNPbin} objects as used
in the \code{genotypes} column of \link{gen_tibble}.
}
\author{
Thibaut Jombart for the original \link[adegenet:glAux]{adegenet::glNA}, modified
by Andrea Manica for 'tidypopgen'
}
