% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snpbin_sums.R
\name{snpbin_list_sums}
\alias{snpbin_list_sums}
\title{Sum of genotypes}
\usage{
snpbin_list_sums(.x, alleles_as_units = TRUE, use_c = FALSE)
}
\arguments{
\item{.x}{a list of \code{\link[adegenet:SNPbin]{adegenet::SNPbin}} objects (usually the \code{genotype} column of
a \code{\link{gen_tibble}} object),}

\item{alleles_as_units}{a logical indicating whether alleles are considered
as units (i.e., a diploid genotype equals two samples, a triploid, three,
etc.) or whether individuals are considered as units of information.}

\item{use_c}{a logical indicating whether compiled C code should be used
(TRUE) or not (FALSE, default).}
}
\value{
a \code{\link[tibble:tibble]{tibble::tibble}} of information (see\code{\link{gen_tibble}} for details
on compulsory columns that will always be present)
}
\description{
This function sums the number of alternate alleles in each SNP. It is not
meant to be used as a general summary, but it is used within other functions to
compute intermediate quantities.
When ploidy varies across individuals, the outputs of this function depend
on whether the information units are individuals, or
alleles within individuals (see details).
}
\details{
=== On the unit of information ===
In the cases where individuals can have different ploidy, computation of
sums, means, etc. of allelic data depends on what we consider as a
unit of information.
To estimate e.g. allele frequencies, unit of information can be
considered as the allele, so that a diploid genotype contains two
samples, a triploid individual, three samples, etc. In such a case,
all computations are done directly on the number of alleles. This
corresponds to allele_as_unit = TRUE.

However, when the focus is put on studying differences/similarities
between individuals, the unit of information is the individual, and all
genotypes possess the same information no matter what their ploidy is.
In this case, computations are made after standardizing individual
genotypes to relative allele frequencies. This corresponds
to allele_as_units = FALSE.

Note that when all individuals have the same ploidy, this distinction
does not hold any more.

This function is a modified version
of \link[adegenet:glAux]{adegenet::glSum}, recoded to work on lists of \code{SNPbin} objects as used
in the \code{genotypes} column of \link{gen_tibble}.
}
\author{
Thibaut Jombart for the original \link[adegenet:glAux]{adegenet::glSum}, modified
by Andrea Manica for 'tidypopgen'
}
