% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snpbin_list_vars.R
\name{snpbin_list_vars}
\alias{snpbin_list_vars}
\title{Compute the variances of loci}
\usage{
snpbin_list_vars(.x, alleles_as_units = TRUE, use_c = FALSE)
}
\arguments{
\item{.x}{a list of \code{\link[adegenet:SNPbin]{adegenet::SNPbin}} objects (usually the \code{genotype} column of
a \code{\link{gen_tibble}} object),}

\item{alleles_as_units}{a logical indicating whether alleles are considered
as units (i.e., a diploid genotype equals two samples, a triploid, three,
etc.) or whether individuals are considered as units of information.}

\item{use_c}{a logical indicating whether compiled C code should be used
(TRUE) or not (FALSE, default).}
}
\value{
a vector of counts of NAs
}
\description{
This function compute the variance for each locus (taking into account
missing genotypes). It is not
meant to be used as a general summary, but it is used within other functions to
compute intermediate quantities.
When ploidy varies across individuals, the outputs of this function depend
on whether the information units are individuals, or
alleles within individuals (see details for \code{\link[=snpbin_list_sums]{snpbin_list_sums()}}).
}
\details{
This function is a modified version
of \link[adegenet:glAux]{adegenet::glNA}, recoded to work on lists of \code{SNPbin} objects as used
in the \code{genotypes} column of \link{gen_tibble}.
}
\author{
Thibaut Jombart for the original \link[adegenet:glAux]{adegenet::glNA}, modified
by Andrea Manica for 'tidypopgen'
}
