% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_tibble.R
\name{stopifnot_gen_tibble}
\alias{stopifnot_gen_tibble}
\title{Test if a tibble is really \code{gen_tibble}}
\usage{
stopifnot_gen_tibble(.x)
}
\arguments{
\item{.x}{the tibble}
}
\description{
Some \code{dplyr} operations strip the subclass from the tibble. This function
is used to check if the tibble is, in reality, still of class \code{gen_tbl}
}
\keyword{internal}
