% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_tibble.R
\name{stopifnot_snpbin_list}
\alias{stopifnot_snpbin_list}
\title{Test if a list is made of \code{SNPbin}}
\usage{
stopifnot_snpbin_list(.x)
}
\arguments{
\item{.x}{the tibble}
}
\description{
At some point, this will be obsolete if we use \code{vctrs}. But for the moment,
it allows us to check if a list is genotypes is properly formatted.
}
\keyword{internal}
