% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{str_replace_base}
\alias{str_replace_base}
\title{Replace matches with new text}
\usage{
str_replace_base(string, pattern, replacement)
}
\arguments{
\item{string}{A character input vector}

\item{pattern}{A regular expression}

\item{replacement}{A character vector}
}
\value{
A character vector with the first match replaced
}
\description{
This function is a drop-in replacement for \code{stringr::str_replace()}, which
replaces the first match
}
\examples{
x <- "I love cats and cats are cute"
str_replace_base(x, "cats", "dogs")
}
\keyword{internal}
