% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_int_names.R
\name{[.matrix_int_names}
\alias{[.matrix_int_names}
\title{Subsetting method for matrix_int_names}
\usage{
\method{[}{matrix_int_names}(x, i, j, drop = TRUE)
}
\description{
Subsetting method for matrix_int_names
}
