% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_bigSNP.R
\name{subset_bigSNP}
\alias{subset_bigSNP}
\title{Subset a bigSNP object}
\usage{
subset_bigSNP(
  X,
  indiv_indices = NULL,
  loci_indices = NULL,
  swap_indices = NULL,
  backingfile = NULL
)
}
\arguments{
\item{X}{the \code{\link[bigsnpr:bigSNP-class]{bigsnpr::bigSNP}} object to subset}

\item{indiv_indices}{the indices of the individuals (rows) to keep}

\item{loci_indices}{the indices of the loci (columns) to keep}

\item{swap_indices}{the indices of loci where the alleles should be swapped
(both in the bim table and in the genotypes)}

\item{backingfile}{the backing file (if null, a tempfile will be used)}
}
\value{
a \code{\link[bigsnpr:bigSNP-class]{bigsnpr::bigSNP}} object
}
\description{
This function subsets an \code{\link[bigsnpr:bigSNP-class]{bigsnpr::bigSNP}} object, creating a new backing
file with the information subsetted and reorders as required.
}
\keyword{internal}
