% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_admix_methods.R
\name{summary.gt_admix}
\alias{summary.gt_admix}
\title{Summary method for gt_admix objects}
\usage{
\method{summary}{gt_admix}(object, ...)
}
\arguments{
\item{object}{a \code{gt_admix} object}

\item{...}{unused (necessary for compatibility with generic function)}
}
\value{
A summary of the \code{gt_admix} object
}
\description{
Summary method for gt_admix objects
}
\examples{
# run the example only if we have the package installed
if (requireNamespace("LEA", quietly = TRUE)) {
  example_gt <- load_example_gt("gen_tbl")

  # Create a gt_admix object
  admix_obj <- example_gt \%>\% gt_snmf(k = 1:3, project = "force")

  # Print a summary
  summary(admix_obj)
}

}
