% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q_matrix.R
\name{summary.q_matrix_list}
\alias{summary.q_matrix_list}
\title{Summarise a Q matrix list}
\usage{
\method{summary}{q_matrix_list}(object, ...)
}
\arguments{
\item{object}{A \code{q_matrix_list} object.}

\item{...}{not currently used}
}
\value{
A summary of the object.
}
\description{
Takes a \code{q_matrix_list} object and returns a summary.
}
