% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_distruct.R
\name{theme_distruct}
\alias{theme_distruct}
\title{A theme to match the output of distruct}
\usage{
theme_distruct()
}
\value{
a \link[ggplot2:theme]{ggplot2::theme}
}
\description{
A theme to remove most plot decorations, matching the look of plots
created with distruct.
}
