% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_pca_tidiers.R
\name{tidy.gt_pca}
\alias{tidy.gt_pca}
\alias{gt_pca_tidiers}
\title{tidy a \code{gt_pca} object}
\usage{
\method{tidy}{gt_pca}(x, matrix = "u", ...)
}
\arguments{
\item{x}{A \code{gt_pca} object returned by one of the \verb{gt_pca_*} functions.}

\item{matrix}{Character specifying which component of the PCA should be
tidied.
\itemize{
\item \code{"u"}, \code{"samples"}, \code{"scores"}, or \code{"x"}: returns information about
the map from the original space into principle components space.
\item \code{"v"}, \code{"rotation"}, \code{"loadings"} or \code{"variables"}: returns information
about the map from principle components space back into the original
space.
\item \code{"d"}, \code{"eigenvalues"} or \code{"pcs"}: returns information about the
eigenvalues.
}}

\item{...}{Not used. Needed to match generic signature only.}
}
\value{
A \link[tibble:tibble]{tibble::tibble} with columns depending on the component of
PCA being tidied.

If \code{"scores"} each row in the
tidied output corresponds to the original data in PCA space. The columns
are:

\item{\code{row}}{ID of the original observation (i.e. rowname from original
data).}
\item{\code{PC}}{Integer indicating a principal component.}
\item{\code{value}}{The score of the observation for that particular principal
component. That is, the location of the observation in PCA space.}

If \code{matrix} is  \code{"loadings"}, each
row in the tidied output corresponds to information about the principle
components in the original space. The columns are:

\item{\code{row}}{The variable labels (colnames) of the data set on
which PCA was performed.}
\item{\code{PC}}{An integer vector indicating the principal component.}
\item{\code{value}}{The value of the eigenvector (axis score) on the
indicated principal component.}

If \code{"eigenvalues"}, the columns are:

\item{\code{PC}}{An integer vector indicating the principal component.}
\item{\code{std.dev}}{Standard deviation explained by this PC.}
\item{\code{percent}}{Fraction of variation explained by this component
(a numeric value between 0 and 1).}
\item{\code{cumulative}}{Cumulative fraction of variation explained by
principle components up to this component (a numeric value between 0 and
1).}
}
\description{
This summarizes information about the components of a \code{gt_pca} from the
\code{tidypopgen} package. The parameter \code{matrix} determines which element is
returned. Column names of the tidied output match those returned by
\link[broom:tidy.prcomp]{broom::tidy.prcomp}, the tidier for the standard PCA objects returned
by \link[stats:prcomp]{stats::prcomp}.
}
\seealso{
\code{\link[=gt_pca_autoSVD]{gt_pca_autoSVD()}} \code{\link[=augment.gt_pca]{augment.gt_pca()}}
}
