% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_dapc_tidiers.R
\name{tidy_q}
\alias{tidy_q}
\title{Tidy a Q matrix}
\usage{
tidy_q(q_mat, gen_tbl)
}
\arguments{
\item{q_mat}{A Q matrix object (as returned by LEA::Q()).}

\item{gen_tbl}{An associated gen_tibble}
}
\value{
A tidied matrix
}
\description{
Tidy a Q matrix
}
