% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidypopgen-package.R
\docType{package}
\name{tidypopgen-package}
\alias{tidypopgen}
\alias{tidypopgen-package}
\title{tidypopgen: Tidy Population Genetics}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

We provide a tidy grammar of population genetics, facilitating the manipulation and analysis of data on biallelic single nucleotide polymorphisms (SNPs). 'tidypopgen' scales to very large genetic datasets by storing genotypes on disk, and performing operations on them in chunks, without ever loading all data in memory. The full functionalities of the package are described in Carter et al. (2025) \doi{10.1111/2041-210x.70204}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/EvolEcolGroup/tidypopgen}
  \item \url{https://evolecolgroup.github.io/tidypopgen/}
  \item Report bugs at \url{https://github.com/EvolEcolGroup/tidypopgen/issues}
}

}
\author{
\strong{Maintainer}: Andrea Manica \email{am315@cam.ac.uk} (\href{https://orcid.org/0000-0003-1895-450X}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Andrea Manica \email{am315@cam.ac.uk} (\href{https://orcid.org/0000-0003-1895-450X}{ORCID}) [copyright holder]
  \item Evie Carter
  \item Eirlys Tysall
}

Other contributors:
\itemize{
  \item Chang Christopher (Author of Hardy-Weinberg Equilibrium algorithm in PLINK 1.90, used in loci_hwe()) [contributor]
  \item Shaun Purcell (Author of Hardy-Weinberg Equilibrium algorithm in PLINK 1.90, used in loci_hwe()) [contributor]
  \item Bengtsson Henrik (Author of countLines in R.utils, modified for .vcf in count_vcf_variants()) [contributor]
}

}
\keyword{internal}
