% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_to_fbm_cpp_new.R
\name{vcf_to_fbm_cpp_new}
\alias{vcf_to_fbm_cpp_new}
\title{Convert vcf to FBM.}
\usage{
vcf_to_fbm_cpp_new(
  vcf_path,
  chunk_size = NULL,
  backingfile = NULL,
  n_cores = 1,
  quiet = FALSE
)
}
\arguments{
\item{vcf_path}{the path to the vcf}

\item{chunk_size}{the chunk size to use on the vcf when loading the file. If
NULL, a best guess will be made.}

\item{backingfile}{the name of the file to use as the backing file for the
FBM. If NULL, the vcf path will be used.}

\item{n_cores}{the number of cores to use when reading the vcf file. Default
is 1.}
}
\value{
path to the resulting rds file as class bigSNP.
}
\description{
Convert a vcf file to a Filebacked Big Matrix (FBM) object. This should work
even for large vcf files that would not fit in memory.
}
\keyword{internal}
