% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_pairwise_pop_pbs.R, R/window_pop_pbs.R
\name{window_pairwise_pop_pbs}
\alias{window_pairwise_pop_pbs}
\title{Compute the Population Branch Statistics over a sliding window}
\usage{
window_pairwise_pop_pbs(
  .x,
  method = "Hudson",
  return_fst = FALSE,
  window_size,
  step_size,
  size_unit = c("snp", "bp"),
  min_loci = 1,
  complete = FALSE
)

window_pairwise_pop_pbs(
  .x,
  method = "Hudson",
  return_fst = FALSE,
  window_size,
  step_size,
  size_unit = c("snp", "bp"),
  min_loci = 1,
  complete = FALSE
)
}
\arguments{
\item{method}{the method to use for calculating Fst. Currently only "Hudson"
is supported.}

\item{return_fst}{a logical value indicating whether to return the Fst values}

\item{window_size}{The size of the window to use for the estimates.}

\item{step_size}{The step size to use for the windows.}

\item{size_unit}{Either "snp" or "bp". If "snp", the window size and step
size are in number of SNPs. If "bp", the window size and step size are in
base pairs.}

\item{min_loci}{The minimum number of loci required to calculate a window
statistic. If the number of loci in a window is less than this, the window
statistic will be NA.}

\item{complete}{Should the function be evaluated on complete windows only? If
FALSE, the default, then partial computations will be allowed at the end of
the chromosome.}

\item{x}{a grouped \code{gen_tibble} object}
}
\value{
a data frame with the following columns:
\itemize{
\item \code{chromosome}: the chromosome for the window
\item \code{start}: the starting locus of the window
\item \code{end}: the ending locus of the window
\item \code{pbs_a.b.c}: the PBS value for population a given b & c (there
will be multiple such columns covering all 3 way combinations of
populations in the grouped \code{gen_tibble} object)
\item \code{fst_a.b}: the Fst value for population a and b, if \code{return_fst} is TRUE
}

a data frame with the following columns:
\itemize{
\item \code{chromosome}: the chromosome for the window
\item \code{start}: the starting locus of the window
\item \code{end}: the ending locus of the window
\item \code{pbs_a.b.c}: the PBS value for population a given b & c (there
will be multiple such columns covering all 3 way combinations of
populations in the grouped \code{gen_tibble} object)
\item \code{fst_a.b}: the Fst value for population a and b, if \code{return_fst} is TRUE
}
}
\description{
This function computes pairwise Fst for a sliding window across
each chromosome.

The function computes the population branch statistics (PBS) for
a sligiding window for each combination of populations at each locus. The
PBS is a measure of the genetic differentiation between one focal
population and two reference populations, and is used to identify outlier
loci that may be under selection.
}
