% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windows_stats_generic.R
\name{window_stats_generic}
\alias{window_stats_generic}
\title{Estimate window statistics from per locus estimates}
\usage{
window_stats_generic(
  x,
  loci_table,
  operator = c("mean", "sum"),
  window_size,
  window_step,
  size_unit = c("snp", "bp"),
  min_loci = 1,
  complete = FALSE
)
}
\arguments{
\item{x}{A vector containing the per locus estimates.}

\item{loci_table}{a dataframe with columns: 'chromosome', 'position', and
'name'.}

\item{operator}{The operator to use for the window statistics. Either
"mean" or "sum".}

\item{window_size}{The size of the window to use for the estimates.}

\item{window_step}{The step size to use for the windows.}

\item{size_unit}{Either "snp" or "bp". If "snp", the window size and step
size are in number of SNPs. If "bp", the window size and step size are in
base pairs.}

\item{min_loci}{The minimum number of loci required to calculate a window
statistic. If the number of loci in a window is less than this, the window
statistic will be NA.}

\item{complete}{Should the function be evaluated on complete windows only?
If FALSE, the default, then partial computations will be allowed at the end
of the chromosome.}
}
\value{
A tibble with columns: 'chromosome', 'start', 'end', 'stats', and
'n_loci'. The 'stats' column contains the mean of the per locus estimates
in the window, and 'n_snps' contains the number of loci in the window.
}
\description{
This is a general function to estimate window statistics from per locus
estimates.
}
