% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_bed_from_dfs.R
\name{write_bed_from_dfs}
\alias{write_bed_from_dfs}
\title{Simple function to write a plink BED file from data.frames}
\usage{
write_bed_from_dfs(genotypes, loci, indiv_meta, path_out = NULL)
}
\arguments{
\item{genotypes}{a matrix of genotype dosages}

\item{loci}{a table of loci information (as used in \code{\link{gen_tibble}})}

\item{indiv_meta}{a table with 'id' and 'population'}

\item{path_out}{is the path and the prefix, to which '.bed', '.bim', and '.fam'
will be added.}
}
\value{
NULL it writes the files
}
\description{
This function write a bed file using information in tables with the same format
as they are stored in various parts of a \code{\link{gen_tibble}}. This is mostly used
for creating small test files, but could also be used to convert small datasets
manually into BED files (and the appropriate BIM and FAM companions).
}
