% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_bed_from_vcf.R
\name{write_bed_from_vcf}
\alias{write_bed_from_vcf}
\title{Convert vcf to bed}
\usage{
write_bed_from_vcf(vcf_path, bed_path = NULL, quiet = TRUE, ...)
}
\arguments{
\item{vcf_path}{the path to the vcf}

\item{bed_path}{the path where the bed file should be saved. If left to null,
the vcf path will be used by switching the file type from .vcf to .bed}

\item{quiet}{if TRUE, don't output information to screen}

\item{...}{parameters to be passed to \link[vcfR:io_vcfR]{vcfR::read.vcfR}}
}
\value{
the file path where the BED file was saved.
}
\description{
This function quickly converts a smallish vcf to bed. The vcf needs to be
read into memory and manipulated in memory, so there are some serious limits
to this function. PLINK is a better option for converting vcf to BED, but
for small datasets, it should do the job.
}
