% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_plink_raw.R
\name{write_plink_raw}
\alias{write_plink_raw}
\title{Export a \code{gen_tibble} object to PLINK raw format}
\usage{
write_plink_raw(x, file, chunk_size = 10000, overwrite = TRUE)
}
\arguments{
\item{x}{a \link{`gen_tibble} object}

\item{file}{a character string giving the path to the file to convert,
with the extension ".raw"}

\item{chunk_size}{very large objects are written in chunks to minimise the
memory footprint. Increase \code{chunk_site} for higher speed, decrease it for
smaller memory footprint (defaults to 10000)}

\item{overwrite}{boolean whether to overwrite the file}
}
\value{
TRUE if successful
}
\description{
This function exports all the information of a \code{gen_tibble} object into
a PLINK raw file.
}
\details{
This function generates a PLINK raw file with alternate allele information
(as outputted using \code{--include-alt} and
\code{--recodeA}), as well as a map file (as generated with \code{--recode}).
}
