% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_q_matrix.R
\name{as_q_matrix}
\alias{as_q_matrix}
\title{Convert a Q matrix into a \code{q_matrix} obejct}
\usage{
as_q_matrix(x)
}
\arguments{
\item{x}{a matrix}
}
\value{
a \code{q_matrix} object, which is a matrix with appropriate column names
(.QX, where X is the component number) to use with plotting
}
\description{
Takes a matrix of Q values, check its validity, and then formats it correctly
to make sure it can then be processed and plotted correctly
}
