% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_dapc.R, R/gt_pca.R, R/gt_pca_find_cluster.R
\name{autoplot.gt_pca}
\alias{autoplot.gt_pca}
\alias{autoplot.gt_pca_clust}
\title{Autoplots for \code{gt_dapc} objects}
\usage{
\method{autoplot}{gt_pca}(object, type = c("screeplot", "scores", "loadings"), k = NULL, ...)

\method{autoplot}{gt_pca}(object, type = c("screeplot", "scores", "loadings"), k = NULL, ...)

\method{autoplot}{gt_pca_clust}(object, metric = c("BIC", "AIC", "WSS"))
}
\arguments{
\item{object}{an object of class \code{gt_dapc}}

\item{type}{the type of plot (one of "screeplot", "scores" and "loadings")}

\item{k}{the principal components to be plotted: for scores, a pair of values
e.g. c(1,2); for \code{loadings} either one or more values.}

\item{...}{not currently used.}

\item{ld}{the principal components to be plotted: for scores, a pair of values
e.g. c(1,2); for \code{loadings} either one or more values.}
}
\value{
a \code{ggplot2} object

a \code{ggplot2} object

a \code{ggplot2} object
}
\description{
For \code{gt_dapc}, the following types of plots are available:
\itemize{
\item \code{screeplot}: a plot of the eigenvalues of the discriminant axes
\item \code{scores} a scatterplot of the scores of each individual on two discriminant
axes (defined by \code{ld})
\item \code{loadings} a plot of loadings of all loci for a discriminant axis (chosen with \code{ld})
}

For \code{gt_pca}, the following types of plots are available:
\itemize{
\item \code{screeplot}: a plot of the eigenvalues of the principal components (currently
it plots the singular value)
\item \code{scores} a scatterplot of the scores of each individual on two principal
components (defined by \code{pc})
\item \code{loadings} a plot of loadings of all loci for a given component (chosen with \code{pc})
}

For \code{gt_pca_clust}, \code{autoplot} produces a plot of a metric of choice ('BIC',
'AIC' or 'WSS') against the number of clusters (\emph{k}). This plot is can be
used to infer the best value of \emph{k}, which corresponds to the smallest
value of the metric (the minimum in an 'elbow' shaped curve). In some cases,
there is not 'elbow' and the metric keeps decreasing with increasing \emph{k};
in such cases, it is customary to choose the value of \emph{k} at which the
decrease in the metric reaches as plateau. For a programmatic way of choosing
\emph{k}, use \code{\link[=gt_pca_clust_best_k]{gt_pca_clust_best_k()}}.
}
\details{
\code{autoplot} produces simple plots to quickly inspect an object. They are
not customisable; we recommend that you use \code{ggplot2} to produce publication
ready plots.

\code{autoplot} produces simple plots to quickly inspect an object. They are
not customisable; we recommend that you use \code{ggplot2} to produce publication
ready plots.

\code{autoplot} produces simple plots to quickly inspect an object. They are
not customisable; we recommend that you use \code{ggplot2} to produce publication
ready plots.
}
