% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot_gt_pcadapt.R
\name{autoplot_gt_pcadapt}
\alias{autoplot_gt_pcadapt}
\alias{autoplot.gt_pcadapt}
\title{Autoplots for \code{gt_pcadapt} objects}
\usage{
\method{autoplot}{gt_pcadapt}(object, type = c("qq", "manhattan"), ...)
}
\arguments{
\item{object}{an object of class \code{gt_pcadapt}}

\item{type}{the type of plot (one of "qq", and "mnahattan")}

\item{...}{further arguments to be passed to \code{\link[bigsnpr:snp_qq]{bigsnpr::snp_qq()}} or
\code{\link[bigsnpr:snp_manhattan]{bigsnpr::snp_manhattan()}}.}
}
\value{
a \code{ggplot2} object
}
\description{
For \code{gt_pcadapt}, the following types of plots are available:
\itemize{
\item \code{qq}: a qunatile-quantile plot of the p-values from \code{pcadapt}
(wrapping \code{\link[bigsnpr:snp_qq]{bigsnpr::snp_qq()}})
\item \code{manhattan} a manhattan plot of the p-values from \code{pcadapt}
(wrapping \code{\link[bigsnpr:snp_manhattan]{bigsnpr::snp_manhattan()}})
}
}
\details{
\code{autoplot} produces simple plots to quickly inspect an object. They are
not customisable; we recommend that you use \code{ggplot2} to produce publication
ready plots.
}
