% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_report_loci.R
\name{autoplot_l_qc_hwe}
\alias{autoplot_l_qc_hwe}
\title{Internal plotting function for hwe}
\usage{
autoplot_l_qc_hwe(object, hwe_p_vertical_line, hwe_p_low_thresh = NULL, ...)
}
\arguments{
\item{object}{an object of class \code{qc_report_loci}}

\item{hwe_p_vertical_line}{the p-value threshold for the vertical line}

\item{hwe_p_low_thresh}{the p-value threshold to subset the loci}

\item{...}{not currently used.}
}
\value{
a \code{ggplot2} object
}
\description{
Internal plotting function for hwe
}
\keyword{internal}
