% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbind_gen_tibble.R
\name{cbind.gen_tbl}
\alias{cbind.gen_tbl}
\title{Combine a gen_tibble to a data.frame or tibble by column}
\usage{
\method{cbind}{gen_tbl}(..., deparse.level = 1)
}
\arguments{
\item{...}{a gen_tibble and a data.frame or tibble}

\item{deparse.level}{an integer controlling the construction of column names.
See \code{\link[base]{cbind}} for details.}
}
\value{
a \code{gen_tibble}
}
\description{
Combine a gen_tibble to a data.frame or tibble by column
}
