% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_int_names.R
\name{col_names}
\alias{col_names}
\alias{col_names.default}
\alias{col_names.matrix_int_names}
\alias{col_names<-}
\alias{col_names<-.default}
\alias{col_names<-.matrix_int_names}
\title{Get or set column names for matrix_int_names}
\usage{
col_names(x)

\method{col_names}{default}(x)

\method{col_names}{matrix_int_names}(x)

col_names(x) <- value

\method{col_names}{default}(x) <- value

\method{col_names}{matrix_int_names}(x) <- value
}
\arguments{
\item{x}{A matrix_int_names object}

\item{value}{a vector of integer or character column names to set, replacing
the current one.}
}
\value{
Integer or character vector of column names, or NULL
}
\description{
This function reliably returns integer names if present, otherwise character
dimnames. Use this instead of colnames() if you need guaranteed dispatch.
}
\seealso{
Other matrix_int_names_functions: 
\code{\link{[.matrix_int_names}()},
\code{\link{as.data.frame.matrix_int_names}()},
\code{\link{matrix_int_names}()},
\code{\link{print.matrix_int_names}()},
\code{\link{row_names}()}
}
\concept{matrix_int_names_functions}
