% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_tibble.R
\name{gen_tibble}
\alias{gen_tibble}
\title{Constructor for a \code{gen_tibble}}
\usage{
gen_tibble(file_path, backingfile = NULL, quiet = FALSE)
}
\arguments{
\item{file_path}{the path to a plink BED file, or a \code{\link[bigsnpr:bigSNP-class]{bigsnpr::bigSNP}} RDS
file (usually created with \code{\link[bigsnpr:snp_readBed]{bigsnpr::snp_readBed()}})}

\item{backingfile}{the backing file used to store the data, if converting from
BED. If not set, the backing file will be saved in the same directory as the
bed file, using the same file name but with a different file type (.bk rather
than .bed)}

\item{quiet}{provide information on the files used to store the data}
}
\value{
an object of the class \code{gen_tbl}.
}
\description{
A \code{gen_tibble} stores genotypes for individuals in a tidy format. DESCRIBE
here the format
}
