% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_tibble.R
\name{gen_tibble}
\alias{gen_tibble}
\title{Constructor for a \code{gen_tibble}}
\usage{
gen_tibble(bigsnp_path)
}
\arguments{
\item{bigsnp_path}{the path to a \code{\link[bigsnpr:bigSNP-class]{bigsnpr::bigSNP}} file created with
\code{\link[bigsnpr:snp_readBed]{bigsnpr::snp_readBed()}}}
}
\value{
an object of the class \code{gen_tbl}.
}
\description{
A \code{gen_tibble} stores genotypes for individuals in a tidy format. DESCRIBE
here the format
}
