% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_tibble.R
\name{gen_tibble}
\alias{gen_tibble}
\title{Constructor for a \code{gen_tibble}}
\usage{
gen_tibble(ind_meta, genotypes, loci, ploidy = NULL)
}
\arguments{
\item{ind_meta}{a list, data.frame or tibble with compulsory columns 'id'
and 'population', plus any additional metadata of interest.}

\item{genotypes}{a matrix of counts of alternative alleles, one row per
individual and a column per locus}

\item{loci}{a data.frame or tibble, with compulsory columns 'name', 'chromosome',
and 'position'}

\item{ploidy}{a vector giving the ploidy of each individual. If 'ploidy' is not
specified, diploids are assumed.}
}
\value{
an object of the class \code{gen_tbl}.
}
\description{
A \code{gen_tibble} stores genotypes for individuals in a tidy format. DESCRIBE
here the format
}
\examples{
test_ind_meta <- data.frame (id=c("a","b","c"),
                             population = c("pop1","pop1","pop2"))
test_genotypes <- rbind(c(1,1,0,1,1,0),
                        c(2,1,0,0,0,0),
                        c(2,2,0,0,1,1))
test_loci <- data.frame(name=paste0("rs",1:6),
                        chromosome=c(1,1,1,1,2,2),
                        position=c(3,5,65,343,23,456),
                        allele_ref = c("a","t","c","g","c","t"),
                        allele_alt = c("t","c", NA,"c","g","a"))
test_gen <- gen_tibble(ind_meta = test_ind_meta,
                       genotypes = test_genotypes,
                       loci = test_loci)
test_gen
}
