% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_tibble.R
\name{gen_tibble}
\alias{gen_tibble}
\alias{gen_tibble.character}
\alias{gen_tibble.matrix}
\title{Constructor for a \code{gen_tibble}}
\usage{
gen_tibble(x, ..., backingfile = NULL, quiet = FALSE)

\method{gen_tibble}{character}(x, ..., backingfile = NULL, quiet = FALSE)

\method{gen_tibble}{matrix}(x, indiv_meta, loci, ..., backingfile = NULL, quiet = FALSE)
}
\arguments{
\item{x}{can be:
\itemize{
\item a string giving the path to a plink BED file, or a \code{\link[bigsnpr:bigSNP-class]{bigsnpr::bigSNP}}
\item a string giving the path to a RDS file storing a \code{bigSNP} object from
the \code{bigsnpr} package (usually created with \code{\link[bigsnpr:snp_readBed]{bigsnpr::snp_readBed()}})
\item a string giving the path to a vcf file. Note that we currently read the whole
vcf in memory with \code{vcfR}, so only smallish vcf can be imported. Only biallelic
SNPs will be considered.
\item a genotype matrix of dosages (0, 1, 2, NA) giing the dosage of the alternate
allele.
}}

\item{...}{if \code{x} is the name of a vcf file, additional arguments
passed to \code{\link[vcfR:io_vcfR]{vcfR::read.vcfR()}}. Otherwise, unused.}

\item{backingfile}{the path, including the file name without extension,
for backing files used to store the data (they will be given a .bk
and .RDS automatically). This is not needed if \code{x} is already an .RDS file.
If \code{x} is a .BED file and \code{backingfile} is left NULL, the backing file will
be saved in the same directory as the
bed file, using the same file name but with a different file type (.bk rather
than .bed). If \code{x} is a genotype matrix and \code{backingfile} is NULL, then a
temporary file will be created (but note that R will delete it at the end of
the session!)}

\item{quiet}{provide information on the files used to store the data}

\item{indiv_meta}{a list, data.frame or tibble with compulsory columns 'id'
and 'population', plus any additional metadata of interest.}

\item{loci}{a data.frame or tibble, with compulsory columns 'name', 'chromosome',
and 'position','genetic_dist', 'allele_ref' and 'allele_alt'}
}
\value{
an object of the class \code{gen_tbl}.
}
\description{
A \code{gen_tibble} stores genotypes for individuals in a tidy format. DESCRIBE
here the format
}
