% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_as_genind.R
\name{gt_as_genind}
\alias{gt_as_genind}
\title{Convert a \code{gen_tibble} to a \code{genind} object from \code{adegenet}}
\usage{
gt_as_genind(x)
}
\arguments{
\item{x}{a \code{\link{gen_tibble}}, with population coded as 'population'}
}
\value{
a \code{genind} object
}
\description{
This function converts a \code{gen_tibble} to a \code{genind} object from \code{adegenet}
}
\examples{
\dontshow{if (rlang::is_installed("adegenet")) withAutoprint(\{ # examplesIf}
example_gt <- load_example_gt("gen_tbl")

# Convert to genind
gt_genind <- example_gt \%>\% gt_as_genind()

# Check object class
class(gt_genind)
\dontshow{\}) # examplesIf}
}
