% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_write_plink.R
\name{gt_as_plink}
\alias{gt_as_plink}
\title{Export a \code{gen_tibble} object to PLINK bed format}
\usage{
gt_as_plink(x, bedfile = NULL, overwrite = TRUE)
}
\arguments{
\item{x}{a \code{\link{gen_tibble}} object}

\item{bedfile}{a character string giving the path to output bed file.}

\item{overwrite}{boolean whether to overwrite the file.}
}
\value{
TRUE if successful
}
\description{
This function exports all the information of a \code{gen_tibble} object into
a PLINK bed file.
}
