% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_as_plink.R
\name{gt_as_plink}
\alias{gt_as_plink}
\title{Export a \code{gen_tibble} object to PLINK bed format}
\usage{
gt_as_plink(
  x,
  file = NULL,
  type = c("bed", "ped", "raw"),
  overwrite = TRUE,
  use_integer = NULL
)
}
\arguments{
\item{x}{a \code{\link{gen_tibble}} object}

\item{file}{a character string giving the path to output file. If left to NULL,
the output file will have the same path and prefix of the backingfile.}

\item{type}{one of "bed", "ped" or "raw"}

\item{overwrite}{boolean whether to overwrite the file.}

\item{use_integer}{boolean whether to use the integer representation of the chromosomes}
}
\value{
the path of the saved file
}
\description{
This function exports all the information of a \code{gen_tibble} object into
a PLINK bed, ped or raw file (and associated files, i.e. .bim and .fam for .bed;
.fam for .ped).
}
