% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_ibs.R
\name{gt_ibs}
\alias{gt_ibs}
\title{Compute the Identity by State Matrix for a \code{gen_tibble} object}
\usage{
gt_ibs(
  x,
  as_counts = FALSE,
  block_size = bigstatsr::block_size(length(show_loci_names(x)))
)
}
\arguments{
\item{x}{a \code{gen_tibble} object.}

\item{as_counts}{whether the counts of similar alleles, rather than the proportion,
should be returned (FALSE by default).}

\item{block_size}{maximum number of loci read at once. More loci should improve speed,
but will tax memory.}
}
\description{
This function computes the IBS matrix. Missing genotypes will have to be
imputed before using this function. Computing the IBS of 1k individuals for
a million markers will take over 10 mins, so the function will not scale to
huge datasets. But for smaller dataset, it allows estimating the IBS directly
in R.
}
\details{
Note that monomorphic sites are currently counted. It would be better to filter
them beforehand.
}
