% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_pca.R
\name{gt_pca}
\alias{gt_pca}
\title{Principal Component Analysis for \code{gen_tibble} objects}
\usage{
gt_pca(x, k = 10, method = c("autoSVD", "randomSVD"), ...)
}
\arguments{
\item{x}{a \code{gen_tbl} object}

\item{k}{the number of principal components to return/compute}

\item{method}{only "autoSVD" is currently implemented.}

\item{...}{further parameters to pass to the appropriate method}
}
\value{
a \code{gt_pca} object, which is a subclass of \code{bigSVD}
}
\description{
This function performs Principal Component Analysis on a \code{gen_tibble}.
It is currently just a wrapper around \code{\link[bigsnpr:snp_autoSVD]{bigsnpr::snp_autoSVD()}}
}
