% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_pca_partialSVD.R
\name{gt_pca_partialSVD}
\alias{gt_pca_partialSVD}
\title{PCA controlling for LD for \code{gen_tibble} objects}
\usage{
gt_pca_partialSVD(x, k = 10, fun_scaling = bigsnpr::snp_scaleBinom())
}
\arguments{
\item{x}{a \code{gen_tbl} object}

\item{k}{Number of singular vectors/values to compute. Default is \code{10}.
\strong{This algorithm should be used to compute a few singular vectors/values.}}

\item{fun_scaling}{Usually this  can be left unset, as it defaults to
\code{\link[bigsnpr:snp_scaleBinom]{bigsnpr::snp_scaleBinom()}}, which is the appropriate function for biallelic SNPs.
Alternatively it is possible to use  custom function
(see \code{\link[bigsnpr:snp_autoSVD]{bigsnpr::snp_autoSVD()}} for details.}
}
\description{
This function performs Principal Component Analysis on a \code{gen_tibble},
using a fast truncated SVD with initial pruning and then iterative removal
of long-range LD regions. This function is a wrapper for \code{\link[bigsnpr:snp_autoSVD]{bigsnpr::snp_autoSVD()}}
}
