% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_write_plink.R
\name{gt_write_plink}
\alias{gt_write_plink}
\title{Export a \code{gen_tibble} object to PLINK ped format}
\usage{
gt_write_plink(
  x,
  file,
  plink_format = c("raw", "ped"),
  chunk_size = 10000,
  overwrite = TRUE
)
}
\arguments{
\item{x}{a \code{\link{gen_tibble}} object}

\item{file}{a character string giving the path to the file to convert,
with the extension ".raw" or "ped"}

\item{plink_format}{character, one of "raw" or "ped"}

\item{chunk_size}{very large objects are written in chunks to minimise the
memory footprint. Increase \code{chunk_site} for higher speed, decrease it for
smaller memory footprint (defaults to 10000)}

\item{overwrite}{boolean whether to overwrite the file}
}
\value{
TRUE if successful
}
\description{
This function exports all the information of a \code{gen_tibble} object into
a PLINK ped file.
}
\details{
This function generates a PLINK raw file with alternate allele information
(as outputted using \code{--include-alt} and
\code{--recodeA}), as well as a map file (as generated with \code{--recode}).
}
