% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_heatmap.R
\name{heatmap_pairwise}
\alias{heatmap_pairwise}
\title{Pairwise distance heatmap}
\usage{
heatmap_pairwise(x, cluster = FALSE, hclust_method = "complete")
}
\arguments{
\item{x}{A square numeric matrix OR a data.frame/tibble with 3 columns:
names1, names2, value (half-matrix representation).}

\item{cluster}{Logical. If TRUE, reorder using hierarchical clustering
(based on \code{\link[stats:hclust]{stats::hclust()}}).}

\item{hclust_method}{Character. The agglomeration method for
\code{\link[stats:hclust]{stats::hclust()}}. It defaults to "complete".}
}
\value{
A ggplot object.
}
\description{
Plot a square heatmap from either a full symmetric distance matrix or a tidy
tibble of distances. It supports optional hierarchical clustering for
row/column reordering using \code{\link[stats:hclust]{stats::hclust()}}.
}
\examples{
example_gt <- load_example_gt("gen_tbl")

# With a pairwise matrix:
fst_dist <- example_gt \%>\%
  group_by(population) \%>\%
  pairwise_pop_fst(method = "Nei87", type = "pairwise")
heatmap_pairwise(fst_dist, cluster = FALSE)
heatmap_pairwise(fst_dist, cluster = TRUE)

# With a tidy tibble of distances:
fst_tidy <- example_gt \%>\%
  group_by(population) \%>\%
  pairwise_pop_fst(method = "Nei87", type = "tidy")
heatmap_pairwise(fst_tidy, cluster = FALSE)
heatmap_pairwise(fst_tidy, cluster = TRUE)

}
