% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heterozygosity.R
\name{heterozygosity}
\alias{heterozygosity}
\alias{heterozygosity.gen_tbl}
\alias{heterozygosity.list}
\title{Estimate individual heterozygosity}
\usage{
heterozygosity(.x, ...)

\method{heterozygosity}{gen_tbl}(.x, .col, ...)

\method{heterozygosity}{list}(.x, ...)
}
\arguments{
\item{.x}{a list of \code{\link[adegenet:SNPbin]{adegenet::SNPbin}} objects (usually the \code{genotype} column of
a \code{\link{gen_tibble}} object),
or a \code{\link{gen_tibble}}.}

\item{...}{currently unused.}

\item{.col}{if \code{.x} is a \code{\link{gen_tibble}}, the column containing the genotypes
(usually \code{genotypes})}
}
\value{
a vector of heterozygosities, one per individuals in the \code{\link{gen_tibble}}
}
\description{
Estimate heterozygosity for each individual.
}
