% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heterozygosity.R
\name{heterozygosity}
\alias{heterozygosity}
\alias{heterozygosity.gen_tbl}
\alias{heterozygosity.list}
\title{Estimate individual heterozygosity}
\usage{
heterozygosity(.data, ...)

\method{heterozygosity}{gen_tbl}(.data, ...)

\method{heterozygosity}{list}(.data, ...)
}
\arguments{
\item{.data}{a \code{\link{gen_tibble}} object.}

\item{...}{currently unused}
}
\value{
a vector of heterozygosities
}
\description{
Estimate heterozygosity for each individual.
}
