% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ind_H_obs.R
\name{ind_H_obs}
\alias{ind_H_obs}
\alias{ind_H_obs.gen_tbl}
\alias{ind_H_obs.list}
\title{Estimate individual observed heterozygosity}
\usage{
ind_H_obs(.x, ...)

\method{ind_H_obs}{gen_tbl}(.x, .col, ...)

\method{ind_H_obs}{list}(.x, ...)
}
\arguments{
\item{.x}{a list of \code{\link[adegenet:SNPbin]{adegenet::SNPbin}} objects (usually the \code{genotype} column of
a \code{\link{gen_tibble}} object),
or a \code{\link{gen_tibble}}.}

\item{...}{currently unused.}

\item{.col}{if \code{.x} is a \code{\link{gen_tibble}}, the column containing the genotypes
(usually \code{genotypes})}
}
\value{
a vector of heterozygosities, one per individuals in the \code{\link{gen_tibble}}
}
\description{
Estimate heterozygosity for each individual.
}
