% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indiv_het_obs.R
\name{indiv_het_obs}
\alias{indiv_het_obs}
\alias{indiv_het_obs.tbl_df}
\alias{indiv_het_obs.vctrs_bigSNP}
\alias{indiv_het_obs.grouped_df}
\title{Estimate individual observed heterozygosity}
\usage{
indiv_het_obs(.x, ...)

\method{indiv_het_obs}{tbl_df}(.x, ...)

\method{indiv_het_obs}{vctrs_bigSNP}(.x, ...)

\method{indiv_het_obs}{grouped_df}(.x, ...)
}
\arguments{
\item{.x}{a vector of class \code{vctrs_bigSNP} (usually the \code{genotype} column of
a \code{\link{gen_tibble}} object),
or a \code{\link{gen_tibble}}.}

\item{...}{currently unused.}
}
\value{
a vector of heterozygosities, one per individuals in the \code{\link{gen_tibble}}
}
\description{
Estimate observed heterozygosity (H_obs) for each individual (i.e. the frequency of
loci that are heterozygous in an individual).
}
