% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indiv_ploidy.R
\name{indiv_ploidy}
\alias{indiv_ploidy}
\alias{indiv_ploidy.tbl_df}
\alias{indiv_ploidy.vctrs_bigSNP}
\title{Return individual ploidy}
\usage{
indiv_ploidy(.x, ...)

\method{indiv_ploidy}{tbl_df}(.x, ...)

\method{indiv_ploidy}{vctrs_bigSNP}(.x, ...)
}
\arguments{
\item{.x}{a \code{\link{gen_tibble}}, or a vector of class \code{vctrs_bigSNP} (usually the
\code{genotype} column of a \code{\link{gen_tibble}} object)}

\item{...}{currently unused.}
}
\value{
a vector of ploidy, one per individuals in the \code{\link{gen_tibble}}
}
\description{
Returns the ploidy for each individual.
}
\examples{
example_gt <- load_example_gt("gen_tbl")

example_gt \%>\% indiv_ploidy()

}
