% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indiv_qc_report.R
\name{indiv_qc_report}
\alias{indiv_qc_report}
\title{Create a Quality Control report for individuals}
\usage{
indiv_qc_report(.x, ...)
}
\arguments{
\item{.x}{a \code{\link{gen_tibble}} object.}

\item{...}{further arguments to pass}
}
\value{
a tibble with 2 elements: het_obs and missingness
}
\description{
#' Return QC information to assess loci (Observed heterozygosity and missingness).
}
