% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loci_alt_freq.R
\name{loci_alt_freq}
\alias{loci_alt_freq}
\alias{loci_alt_freq.tbl_df}
\alias{loci_alt_freq.vctrs_bigSNP}
\alias{loci_alt_freq.grouped_df}
\alias{loci_maf}
\alias{loci_maf.tbl_df}
\alias{loci_maf.vctrs_bigSNP}
\alias{loci_maf.grouped_df}
\title{Estimate allele frequencies at each each locus}
\usage{
loci_alt_freq(.x, n_cores, block_size, type, ...)

\method{loci_alt_freq}{tbl_df}(
  .x,
  n_cores = bigstatsr::nb_cores(),
  block_size = bigstatsr::block_size(nrow(.x), 1),
  ...
)

\method{loci_alt_freq}{vctrs_bigSNP}(
  .x,
  n_cores = bigstatsr::nb_cores(),
  block_size = bigstatsr::block_size(length(.x), 1),
  ...
)

\method{loci_alt_freq}{grouped_df}(
  .x,
  n_cores = bigstatsr::nb_cores(),
  block_size = bigstatsr::block_size(nrow(.x), 1),
  type = "tidy",
  ...
)

loci_maf(.x, n_cores, block_size, ...)

\method{loci_maf}{tbl_df}(
  .x,
  n_cores = bigstatsr::nb_cores(),
  block_size = bigstatsr::block_size(nrow(.x), 1),
  ...
)

\method{loci_maf}{vctrs_bigSNP}(
  .x,
  n_cores = bigstatsr::nb_cores(),
  block_size = bigstatsr::block_size(length(.x), 1),
  ...
)

\method{loci_maf}{grouped_df}(
  .x,
  n_cores = bigstatsr::nb_cores(),
  block_size = bigstatsr::block_size(nrow(.x), 1),
  ...
)
}
\arguments{
\item{.x}{a vector of class \code{vctrs_bigSNP} (usually the \code{genotypes} column of
a \code{\link{gen_tibble}} object), or a \code{\link{gen_tibble}}.}

\item{n_cores}{number of cores to be used, it defaults to
\code{\link[bigstatsr:reexports]{bigstatsr::nb_cores()}}}

\item{block_size}{maximum number of loci read at once.}

\item{type}{type of object to return, if using grouped method. One of "tidy",
"list", or "matrix". Default is "tidy".}

\item{...}{other arguments passed to specific methods, currently unused.}
}
\value{
a vector of frequencies, one per locus
}
\description{
Allele frequencies can be estimates as minimum allele frequencies (MAF) with
\code{loci_maf()} or the frequency of the alternate allele (with
\code{loci_alt_freq()}). The latter are in line with the genotypes matrix (e.g. as
extracted by \code{\link[=show_loci]{show_loci()}}). Most users will be in interested in the MAF,
but the raw frequencies might be useful when computing aggregated statistics.
}
