% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loci_freq.R
\name{loci_freq}
\alias{loci_freq}
\alias{loci_freq.tbl_df}
\alias{loci_freq.vctrs_bigSNP}
\alias{loci_freq.grouped_df}
\title{Estimate allele frequencies at each each locus}
\usage{
loci_freq(.x, ...)

\method{loci_freq}{tbl_df}(.x, ..., minor = TRUE)

\method{loci_freq}{vctrs_bigSNP}(.x, ..., minor = TRUE)

\method{loci_freq}{grouped_df}(.x, ..., minor = TRUE)
}
\arguments{
\item{.x}{a vector of class \code{vctrs_bigSNP} (usually the \code{genotypes} column of
a \code{\link{gen_tibble}} object),
or a \code{\link{gen_tibble}}.}

\item{...}{other arguments passed to specific methods.}

\item{minor}{a logical indicating whether we should give the frequencies of
the minor allele (TRUE, the default). If FALSE, the frequencies of the
alternate allele are given.}
}
\value{
a vector of frequencies, one per locus
}
\description{
Estimate the frequency of the alternate allele at each locus.
}
