% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loci_hwe.R
\name{loci_hwe}
\alias{loci_hwe}
\alias{loci_hwe.tbl_df}
\alias{loci_hwe.vctrs_bigSNP}
\alias{loci_hwe.grouped_df}
\title{Test Hardy-Weinberg equilibrium at each locus}
\usage{
loci_hwe(.x, ...)

\method{loci_hwe}{tbl_df}(.x, ...)

\method{loci_hwe}{vctrs_bigSNP}(.x, ...)

\method{loci_hwe}{grouped_df}(.x, ...)
}
\arguments{
\item{.x}{a vector of class \code{vctrs_bigSNP} (usually the \code{genotypes} column of
a \code{\link{gen_tibble}} object),
or a \code{\link{gen_tibble}}.}

\item{...}{further arguments to pass to \code{\link[HardyWeinberg:HWExact]{HardyWeinberg::HWExact()}}.}
}
\value{
a vector of probabilities from HWE exact test, one per locus
}
\description{
Return the p-value from an exact test of HWE. It uses an implementation
based on \code{\link[HardyWeinberg:HWExact]{HardyWeinberg::HWExact()}}), optimised for speed.
}
\details{
NOTE There are no tests for this function yet! Unit tests are needed.
}
\author{
based on code originally written by Jan Graffleman, optimised by Andrea Manica for speed
}
