% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loci_hwe.R
\name{loci_hwe}
\alias{loci_hwe}
\alias{loci_hwe.tbl_df}
\alias{loci_hwe.vctrs_bigSNP}
\title{Test Hardy-Weinberg equilibrium at each locus}
\usage{
loci_hwe(.x, ...)

\method{loci_hwe}{tbl_df}(.x, mid_p = TRUE, ...)

\method{loci_hwe}{vctrs_bigSNP}(.x, mid_p = TRUE, ...)
}
\arguments{
\item{.x}{a vector of class \code{vctrs_bigSNP} (usually the \code{genotypes} column of
a \code{\link{gen_tibble}} object), or a \code{\link{gen_tibble}}.}

\item{...}{not used.}

\item{mid_p}{boolean on whether the mid-p value should be computed. Default
is TRUE, as in PLINK.}
}
\value{
a vector of probabilities from HWE exact test, one per locus
}
\description{
Return the p-value from an exact test of HWE.
}
\details{
This function uses the original C++ algorithm from PLINK 1.90.
}
\examples{
example_gt <- example_gt()

# For HWE
example_gt \%>\% loci_hwe()

# For loci_hwe per locus per population, use reframe
example_gt \%>\%
  group_by(population) \%>\%
  reframe(loci_hwe = loci_hwe(genotypes))

}
\author{
the C++ algorithm was written by Christopher Chang for PLINK 1.90,
based on original code by Jan Wigginton (the code was released under GPL3).
}
