% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loci_missingness.R
\name{loci_missingness}
\alias{loci_missingness}
\alias{loci_missingness.tbl_df}
\alias{loci_missingness.vctrs_bigSNP}
\alias{loci_missingness.grouped_df}
\title{Estimate missingness at each locus}
\usage{
loci_missingness(.x, as_counts = FALSE, block_size, ...)

\method{loci_missingness}{tbl_df}(
  .x,
  as_counts = FALSE,
  block_size = bigstatsr::block_size(nrow(attr(.x$genotypes, "loci")), 1),
  ...
)

\method{loci_missingness}{vctrs_bigSNP}(
  .x,
  as_counts = FALSE,
  block_size = bigstatsr::block_size(nrow(attr(.x, "loci")), 1),
  ...
)

\method{loci_missingness}{grouped_df}(
  .x,
  as_counts = FALSE,
  block_size = bigstatsr::block_size(nrow(attr(.x, "loci")), 1),
  n_cores = bigstatsr::nb_cores(),
  ...
)
}
\arguments{
\item{.x}{a vector of class \code{vctrs_bigSNP} (usually the \code{genotypes} column of
a \code{\link{gen_tibble}} object), or a \code{\link{gen_tibble}}.}

\item{as_counts}{boolean defining whether the count of NAs (rather than the
rate) should be returned. It defaults to FALSE (i.e. rates are returned by
default).}

\item{block_size}{maximum number of loci read at once.}

\item{...}{other arguments passed to specific methods.}

\item{n_cores}{number of cores to be used, it defaults to
\code{\link[bigstatsr:reexports]{bigstatsr::nb_cores()}}}
}
\value{
a vector of frequencies, one per locus
}
\description{
Estimate the rate of missingness at each locus.
}
