% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loci_missingness.R
\name{loci_missingness}
\alias{loci_missingness}
\alias{loci_missingness.tbl_df}
\alias{loci_missingness.vctrs_bigSNP}
\alias{loci_missingness.grouped_df}
\title{Estimate missingness at each locus}
\usage{
loci_missingness(.x, as_counts = FALSE, ...)

\method{loci_missingness}{tbl_df}(.x, as_counts = FALSE, ...)

\method{loci_missingness}{vctrs_bigSNP}(.x, as_counts = FALSE, ...)

\method{loci_missingness}{grouped_df}(.x, as_counts = FALSE, ...)
}
\arguments{
\item{.x}{a list of \code{\link[adegenet:SNPbin]{adegenet::SNPbin}} objects (usually the \code{genotype} column of
a \code{\link{gen_tibble}} object),
or a \code{\link{gen_tibble}}.}

\item{as_counts}{booelean defining whether the count of NAs (rather than the rate)
should be returned. It defaults to FALSE (i.e. rates are returned by default).}

\item{...}{other arguments passed to specific methods.}
}
\value{
a vector of frequencies, one per locus
}
\description{
Estimate the rate of missingness at each locus.
}
