% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loci_pi.R
\name{loci_pi}
\alias{loci_pi}
\alias{loci_pi.tbl_df}
\alias{loci_pi.vctrs_bigSNP}
\alias{loci_pi.grouped_df}
\title{Estimate nucleotide diversity (pi) at each locus}
\usage{
loci_pi(.x, n_cores, block_size, ...)

\method{loci_pi}{tbl_df}(
  .x,
  n_cores = bigstatsr::nb_cores(),
  block_size = bigstatsr::block_size(nrow(.x), 1),
  ...
)

\method{loci_pi}{vctrs_bigSNP}(
  .x,
  n_cores = bigstatsr::nb_cores(),
  block_size = bigstatsr::block_size(length(.x), 1),
  ...
)

\method{loci_pi}{grouped_df}(
  .x,
  n_cores = bigstatsr::nb_cores(),
  block_size = bigstatsr::block_size(nrow(.x), 1),
  ...
)
}
\arguments{
\item{.x}{a vector of class \code{vctrs_bigSNP} (usually the \code{genotypes} column of
a \code{\link{gen_tibble}} object), or a \code{\link{gen_tibble}}.}

\item{n_cores}{number of cores to be used, it defaults to
\code{\link[bigstatsr:reexports]{bigstatsr::nb_cores()}}}

\item{block_size}{maximum number of loci read at once.}

\item{...}{other arguments passed to specific methods, currently unused.}
}
\value{
a vector of frequencies, one per locus
}
\description{
Estimate nucleotide diversity (pi) at each locus, accounting for missing
values. This uses the formula: c_0 * c_1 / (n * (n-1) / 2)
}
