% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loci_sums.R
\name{loci_sums}
\alias{loci_sums}
\alias{loci_sums.tbl_df}
\alias{loci_sums.vctrs_bigSNP}
\alias{loci_sums.grouped_df}
\title{Estimates the sum of genotypes at each each locus}
\usage{
loci_sums(.x, ...)

\method{loci_sums}{tbl_df}(.x, ...)

\method{loci_sums}{vctrs_bigSNP}(.x, ...)

\method{loci_sums}{grouped_df}(.x, ...)
}
\arguments{
\item{.x}{a list of \code{\link[adegenet:SNPbin]{adegenet::SNPbin}} objects (usually the \code{genotype} column of
a \code{\link{gen_tibble}} object),
or a \code{\link{gen_tibble}}.}

\item{...}{other arguments passed to specific methods.}
}
\value{
a vector of frequencies, one per locus
}
\description{
Estimate the sum of the alternate allele at each locus. This is unlikely to be useful
directly, but it is used by other functions that compute various statistics.
}
