% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loci_transversions.R
\name{loci_transversions}
\alias{loci_transversions}
\alias{loci_transversions.tbl_df}
\alias{loci_transversions.vctrs_bigSNP}
\title{Find transversions}
\usage{
loci_transversions(.x, ...)

\method{loci_transversions}{tbl_df}(.x, ...)

\method{loci_transversions}{vctrs_bigSNP}(.x, ...)
}
\arguments{
\item{.x}{a vector of class \code{vctrs_bigSNP} (usually the \code{genotype} column of
a \code{\link{gen_tibble}} object),
or a \code{\link{gen_tibble}}.}

\item{...}{other arguments passed to specific methods.}
}
\value{
a logical vector defining which loci are transversions
}
\description{
Use the loci table to define which loci are transversions
}
\examples{
example_gt <- example_gen_tibble()
example_gt \%>\% loci_transversions()
}
