% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_test_bed.R
\name{make_test_bed}
\alias{make_test_bed}
\title{simple function to write a test plink BED file}
\usage{
make_test_bed(genotypes, loci, ind_meta, path_out)
}
\description{
We start with data in the format that we expect to come out later on
}
