% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_int_names.R
\name{matrix_int_names}
\alias{matrix_int_names}
\title{Create a matrix with integer or character names}
\usage{
matrix_int_names(data, row_names = NULL, col_names = NULL)
}
\arguments{
\item{data}{Matrix data or object coercible to matrix}

\item{row_names}{Integer vector (stored as int_rownames) or character vector
(stored as dimnames)}

\item{col_names}{Integer vector (stored as int_colnames) or character vector
(stored as dimnames)}
}
\value{
A matrix_int_names object
}
\description{
Create a matrix with integer or character names
}
