% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_int_names.R
\name{matrix_int_names}
\alias{matrix_int_names}
\title{Matrix with integer row/column names class}
\usage{
matrix_int_names(data, row_names = NULL, col_names = NULL)
}
\arguments{
\item{data}{Matrix data or object coercible to matrix}

\item{row_names}{Integer vector (stored as int_rownames) or character vector
(stored as dimnames)}

\item{col_names}{Integer vector (stored as int_colnames) or character vector
(stored as dimnames)}
}
\value{
A \code{matrix_int_names} object
}
\description{
Create a matrix with integer or character names for rows and columns. Integer
names are useful when dealing with very large numbers of rows or columns, as
the integer names can be stored more efficiently than character names. Whilst
this is possible for \code{data.frames}, the native \code{matrix} class only allows for
character row and column names via the \code{dimnames} attribute. Use
\code{\link[=row_names]{row_names()}} and \code{\link[=col_names]{col_names()}} to get or set the integer names, which are
stored in special attributes (\code{int_rownames} and \code{int_colnames}) on the
matrix.
}
\details{
This class allows you to have integer names as attributes while still being a
matrix, and provides methods for getting and setting these names. When you
create a \code{matrix_int_names} object, you can provide integer or character
vectors for row and column names. If you provide integer vectors, they will
be stored in special attributes (\code{int_rownames} and \code{int_colnames}) instead
of the standard \code{dimnames}. If you provide character vectors, they will be
stored in the standard \code{dimnames} as usual. You can also mix and match,
having integer names for rows and character names for columns, or vice versa.
Note that, since the row and column names are stored in special attributes,
you have to use \code{\link[=row_names]{row_names()}} and \code{\link[=col_names]{col_names()}} to get or set them, rather
than \code{rownames()} and \code{colnames()}, which will only return character names if
present.
}
\examples{
# Create a matrix with integer row and column names
my_mat <- matrix_int_names(matrix(1:6, nrow = 2),
  row_names = c(10L, 20L),
  col_names = c(100L, 200L, 300L)
)
row_names(my_mat) # returns integer row names
col_names(my_mat) # returns integer column names
}
\seealso{
Other matrix_int_names_functions: 
\code{\link{[.matrix_int_names}()},
\code{\link{as.data.frame.matrix_int_names}()},
\code{\link{col_names}()},
\code{\link{print.matrix_int_names}()},
\code{\link{row_names}()}
}
\concept{matrix_int_names_functions}
