% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_tibble.R
\name{new_vctrs_bigsnp}
\alias{new_vctrs_bigsnp}
\title{create a vctrs_bigSNP}
\usage{
new_vctrs_bigsnp(bigsnp_obj, bigsnp_file, indiv_id)
}
\arguments{
\item{bigsnp_obj}{the bigsnp object}

\item{bigsnp_file}{the file to which the bigsnp object was saved}

\item{indiv_id}{ids of individuals}
}
\value{
a vctrs_bigSNP object
}
\description{
create a vctrs_bigSNP
}
\keyword{internal}
