% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_ibs.R
\name{pairwise_ibs}
\alias{pairwise_ibs}
\title{Compute the Identity by State Matrix for a \code{gen_tibble} object}
\usage{
pairwise_ibs(
  x,
  as_matrix = FALSE,
  type = c("proportion", "adjusted_counts", "raw_counts"),
  block_size = bigstatsr::block_size(count_loci(x))
)
}
\arguments{
\item{x}{a \code{gen_tibble} object.}

\item{as_matrix}{boolean, determining whether the results should be a square symmetrical matrix (TRUE),
or a tidied tibble (FALSE, the default)}

\item{type}{one of "proportion" (equivalent to "ibs" in PLINK), "adjusted_counts" ("distance" in PLINK),
and "raw_counts" (the counts of identical alleles and non-missing alleles, from which the two other quantities are computed)}

\item{block_size}{maximum number of loci read at once. More loci should improve speed,
but will tax memory.}
}
\value{
a \link[bigstatsr:FBM-class]{bigstatsr::FBM} of proportion or adjusted counts, or a list of
two \link[bigstatsr:FBM-class]{bigstatsr::FBM} matrices, one of counts of IBS by alleles,
and one of number of valid alleles (i.e. 2\emph{n_loci - 2}missing_loci)
}
\description{
This function computes the IBS matrix.
}
\details{
Note that monomorphic sites are currently counted. Should we filter
them beforehand? What does plink do?
}
