% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_pop_fst.R
\name{pairwise_pop_fst}
\alias{pairwise_pop_fst}
\title{Compute pairwise population Fst}
\usage{
pairwise_pop_fst(.x, by_locus = FALSE, method = "Hudson")
}
\arguments{
\item{.x}{a grouped \code{\link{gen_tibble}} (as obtained by using \code{\link[dplyr:group_by]{dplyr::group_by()}})}

\item{by_locus}{boolean, determining whether Fst should be returned by locus(TRUE),
or as a single genome wide value obtained by taking the ratio of the mean numerator
and denominator (FALSE, the default).}

\item{method}{currently only "Hudson" is implemented}
}
\value{
a tibble with each pairwise combination as a row (if \code{tidy=TRUE}), else
a matrix
}
\description{
This function computes pairwise Fst, currently using the formulation by
Hudson (following the recommendations by Bhatia et al (2013)).
}
\references{
Bhatia G, Patterson N, Sankararaman S, Price AL. Estimating and Interpreting
FST: The Impact of Rare Variants. Genome Research. 2013;23(9):1514–1521.
}
