% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_fis.R
\name{pop_fis}
\alias{pop_fis}
\title{Compute population specific fis}
\usage{
pop_fis(.x, include_global = FALSE, as_matrix = NULL)
}
\arguments{
\item{.x}{a grouped \code{\link{gen_tibble}} (as obtained by using \code{\link[dplyr:group_by]{dplyr::group_by()}})}

\item{include_global}{boolean determing whether, besides the population specific fis, a global
fis should be appended. Note that this will return a vector of n populations plus 1 (the global value)}

\item{as_matrix}{optional, the matrix of Allele Sharing returned by
\code{\link[=pairwise_as]{pairwise_as()}}. As a number of statistics can be derived from the AS matrix,
it it sometimes more efficient to precompute this matrix.}
}
\value{
a vector of population specific fis (plus the global value if \code{include_global=TRUE})
}
\description{
This function computes population specific fis (as computed by \code{\link[hierfstat:fs.dosage]{hierfstat::fis.dosage()}}).
}
