% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_global_stats.R
\name{pop_global_stats}
\alias{pop_global_stats}
\title{Compute basic population global statistics}
\usage{
pop_global_stats(.x, by_locus = FALSE, n_cores = bigstatsr::nb_cores())
}
\arguments{
\item{.x}{a \code{\link{gen_tibble}} (usually grouped, as obtained by using \code{\link[dplyr:group_by]{dplyr::group_by()}}; use on
a single population will return a number of quantities as NA/NaN)}

\item{by_locus}{boolean, determining whether the statistics should be returned by locus(TRUE),
or as a single genome wide value (FALSE, the default).}

\item{n_cores}{number of cores to be used, it defaults to \code{\link[bigstatsr:reexports]{bigstatsr::nb_cores()}}}
}
\value{
a tibble of population statistics, with populations as rows and statistics as columns
}
\description{
This function computes basic population global statistics, following the formulas used in \code{hietfstat::basic.stats}.
}
\details{
TO BE CORRECTED (THIS IS INCOMPLETE, AND FORMULAE WERE FILLED IN BY COPILOT)
The statistics computed are:
\itemize{
\item observed heterozygosity (\code{Ho})
\item expected heterozygosity (\code{Hs})
\item total heterozygosity (\code{Ht})
\item within population diversity (\code{Dst})
\item total population diversity (\code{Htp})
\item total population diversity (\code{Dstp})
\item Fst
\item Fst prime
\item Fis
\item Dest
}

The formulas used are:
\eqn{Ho= 1-\sum_k \sum_i  P_{k,ii}/n_p}
where \eqn{P_{k,ii}} represents the proportion of homozygote \eqn{i} in sample \eqn{k} and
\eqn{n_{p}} the number of samples.
\eqn{Hs=\tilde{n}/(\tilde{n}-1)[1-\sum_i\bar{p_i^2}-Ho/2\tilde{n}]}
where \eqn{\tilde{n}=np/\sum_k 1/n_k} and
\eqn{\bar{p_i^2}=\sum_k p_{ki}^2/np}
\eqn{Ht=1-\sum_k \bar{p_k^2} + \bar{Hs}/\tilde{n}/np - Ho/2/\tilde{n}/np}
\eqn{Dst=Ht - Hs}
}
