% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_global_stats.R
\name{pop_global_stats}
\alias{pop_global_stats}
\title{Compute basic population global statistics}
\usage{
pop_global_stats(.x, by_locus = FALSE, n_cores = bigstatsr::nb_cores())
}
\arguments{
\item{.x}{a \code{\link{gen_tibble}} (usually grouped, as obtained by using \code{\link[dplyr:group_by]{dplyr::group_by()}}; use on
a single population will return a number of quantities as NA/NaN)}

\item{by_locus}{boolean, determining whether the statistics should be returned by locus(TRUE),
or as a single genome wide value (FALSE, the default).}

\item{n_cores}{number of cores to be used, it defaults to \code{\link[bigstatsr:reexports]{bigstatsr::nb_cores()}}}
}
\value{
a tibble of population statistics, with populations as rows and statistics as columns
}
\description{
This function computes basic population global statistics, following the notation in Nei 1987 (which in turn is based on Nei and Chesser 1983):
\itemize{
\item observed heterozygosity (\eqn{h_o}, column header \code{Ho})
\item expected heterozygosity, also known as gene diversity (\eqn{h_s}, \code{Hs})
\item total heterozygosity (\eqn{h_t}, \code{Ht})
\item within population diversity (\code{Dst})
\item total population diversity (\code{Htp})
\item total population diversity (\code{Dstp})
\item Fst
\item Fst prime
\item Fis
\item Dest
}
}
\details{
We use the notation of Nei 1987. That notation was for loci with \eqn{m} alleles, but in our case we only have two alleles, so \code{m=2}.
\itemize{
\item Within population observed heterozygosity \eqn{\hat{h}_o} for a locus with \eqn{m} alleles is defined as:\cr
\eqn{\hat{h}_o= 1-\sum_{k=1}^{s} \sum_{i=1}^{m}  \hat{X}_{kii}/s}\cr
where\cr
\eqn{\hat{X}_{kii}} represents the proportion of homozygote \eqn{i} in the sample for the \eqn{k}th population and\cr
\eqn{s} the number of populations,\cr
following equation 7.38 in Nei(1987) on pp.164.\cr
\item Within population expected heterozygosity (gene diversity)  \eqn{\hat{h}_s} for a locus with \eqn{m} alleles is defined as:\cr
\eqn{\hat{h}_s=(\tilde{n}/(\tilde{n}-1))[1-\sum_{i=1}^{m}\bar{\hat{x}_i^2}-\hat{h}_o/2\tilde{n}]}\cr
where \cr \eqn{\tilde{n}=s/\sum_k 1/n_k} (i.e the harmonic mean of \eqn{n_k}) and\cr
\eqn{\bar{\hat{x}_i^2}=\sum_k \hat{x}_{ki}^2/s}\cr
following equation 7.39 in Nei(1987) on pp.164.
\item Total heterozygosity (total gene diversity) \eqn{\hat{h}_t} for a locus with \eqn{m} alleles is defined as:\cr
\eqn{\hat{h}_t = 1-\sum_{i=1}^{m} \bar{\hat{x}_i^2} + \hat{h}_s/(\tilde{n}s) - \hat{h}_o/(2\tilde{n}s)}\cr
where \cr
\eqn{\hat{x}_i=\sum_k \hat{x}_{ki}/s}\cr
following equation 7.40 in Nei(1987) on pp.164.\cr
\item The amount of gene diversity among samples \eqn{Dst} is defined as:\cr
\eqn{Dst=Ht - Hs}\cr
\item Total corrected heterozygosity (total gene diversity) \eqn{\hat{h}_t}
}
}
