% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_het_exp.R
\name{pop_het_exp}
\alias{pop_het_exp}
\alias{pop_gene_div}
\title{Compute the population expected heterozygosity}
\usage{
pop_het_exp(
  .x,
  by_locus = FALSE,
  include_global = FALSE,
  n_cores = bigstatsr::nb_cores()
)

pop_gene_div(...)
}
\arguments{
\item{.x}{a \code{\link{gen_tibble}} (usually grouped, as obtained by using \code{\link[dplyr:group_by]{dplyr::group_by()}}, otherwise the full tibble
will be considered as belonging to a single population).}

\item{by_locus}{boolean, determining whether Hs should be returned by locus(TRUE),
or as a single genome wide value (FALSE, the default).}

\item{include_global}{boolean determining whether, besides the population specific estiamtes, a global
estimate should be appended. Note that this will return a vector of n populations plus 1 (the global value),
or a matrix with n+1 columns if \code{by_locus=TRUE}.}

\item{n_cores}{number of cores to be used, it defaults to \code{\link[bigstatsr:reexports]{bigstatsr::nb_cores()}}}
}
\value{
a vector of mean population observed heterozygosities (if \code{by_locus=FALSE}), or a matrix of
estimates by locus (rows are loci, columns are populations, \code{by_locus=TRUE})
}
\description{
This function computes expected population heterozygosity (also
referred to as gene diversity, to avoid the potentially misleading use of the term "expected" in this context), using the formula of Nei (1987).
}
\details{
Within population expected heterozygosity (gene diversity) is defined as:
\eqn{Hs=\tilde{n}/(\tilde{n}-1)[1-\sum_i\bar{p_i^2}-Ho/2\tilde{n}],}

where \eqn{\tilde{n}=np/\sum_k 1/n_k} and
\eqn{\bar{p_i^2}=\sum_k p_{ki}^2/np}
following Nei(1987) in pp.164--5
}
\references{
Nei M. (1987) Molecular Evolutionary Genetics. Columbia University Press
}
