% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_het_obs.R
\name{pop_het_obs}
\alias{pop_het_obs}
\title{Compute the population observed heterozygosity}
\usage{
pop_het_obs(
  .x,
  by_locus = FALSE,
  include_global = FALSE,
  n_cores = bigstatsr::nb_cores()
)
}
\arguments{
\item{.x}{a \code{\link{gen_tibble}} (usually grouped, as obtained by using \code{\link[dplyr:group_by]{dplyr::group_by()}}, otherwise the full tibble
will be considered as belonging to a single population).}

\item{by_locus}{boolean, determining whether Ho should be returned by locus(TRUE),
or as a single genome wide value (FALSE, the default).}

\item{include_global}{boolean determining whether, besides the population specific estiamtes, a global
estimate should be appended. Note that this will return a vector of n populations plus 1 (the global value),
or a matrix with n+1 columns if \code{by_locus=TRUE}.}

\item{n_cores}{number of cores to be used, it defaults to \code{\link[bigstatsr:reexports]{bigstatsr::nb_cores()}}}
}
\value{
a vector of mean population observed heterozygosities (if \code{by_locus=FALSE}), or a matrix of
estimates by locus (rows are loci, clumns are populations, \code{by_locus=TRUE})
}
\description{
This function computes population heterozygosity, using the formula of Nei (1987).
}
\details{
Within population observed heterozygosity is defined as:
\eqn{Ho= 1-\sum_k \sum_i  P_{k,ii}/n_p}
where \eqn{P_{k,ii}} represents the proportion of homozygote \eqn{i} in sample \eqn{k} and
\eqn{n_{p}} the number of samples.
following Nei(1987) in pp.164--5
}
\references{
Nei M. (1987) Molecular Evolutionary Genetics. Columbia University Press
}
