% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_gt_pca.R
\name{predict.gt_pca}
\alias{predict.gt_pca}
\title{Predict scores of a PCA}
\usage{
\method{predict}{gt_pca}(
  object,
  new_data = NULL,
  project_method = c("none", "simple", "OADP", "least_squares"),
  lsq_pcs = c(1, 2),
  block_size = NULL,
  n_cores = 1,
  ...
)
}
\arguments{
\item{object}{the \code{\link{gt_pca}} object}

\item{new_data}{a gen_tibble if scores are requested for a new dataset}

\item{project_method}{a string taking the value of either "simple",
"OADP" (Online Augmentation, Decomposition, and Procrustes (OADP) projection),
or "least_squares" (as done by SMARTPCA)}

\item{lsq_pcs}{a vector of length two with the values of the two principal components
to use for the least square fitting. Only relevant if\code{project_method = 'least_squares'}}

\item{block_size}{number of loci read simultaneously (larger values will speed up
computation, but require more memory)}

\item{n_cores}{number of cores}

\item{...}{no used}
}
\value{
a matrix of predictions, with samples as rows and components as columns. The number
of components depends on how many were estimated in the \code{\link{gt_pca}} object.
}
\description{
Predict the PCA scores for a \code{\link{gt_pca}}, either for the original data or
projecting new data.
}
\references{
Zhang et al (2020). Fast and robust ancestry prediction using
principal component analysis  36(11): 3439–3446.
}
