% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_gt_pca.R
\name{predict.gt_pca}
\alias{predict.gt_pca}
\title{Predict scores of a PCA}
\usage{
\method{predict}{gt_pca}(object, new_data = NULL, block_size = NULL, ...)
}
\arguments{
\item{object}{the \code{\link{gt_pca}} object}

\item{new_data}{a gen_tibble if scores are requested for a new dataset}

\item{block_size}{number of loci read simultaneously (larger values will speed up
computation, but require more memory)}
}
\value{
a matrix of predictions, with samples as rows and components as columns. The number
of components depends on how many were estimated in the \code{\link{gt_pca}} object
}
\description{
Predict the PCA scores for a \code{\link{gt_pca}}, either for the original data or for new data.
}
