% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_int_names.R
\name{print.matrix_int_names}
\alias{print.matrix_int_names}
\title{Print method for matrix_int_names}
\usage{
\method{print}{matrix_int_names}(x, ...)
}
\description{
Print method for matrix_int_names
}
