% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_int_names.R
\name{print.matrix_int_names}
\alias{print.matrix_int_names}
\title{Print method for matrix_int_names}
\usage{
\method{print}{matrix_int_names}(x, ..., n = 10L)
}
\arguments{
\item{x}{A matrix_int_names object}

\item{...}{Additional arguments passed to print}

\item{n}{Optional number of rows/columns to print (default 10)}
}
\value{
Invisibly returns the original object
}
\description{
Print method for matrix_int_names
}
\seealso{
Other matrix_int_names_functions: 
\code{\link{[.matrix_int_names}()},
\code{\link{as.data.frame.matrix_int_names}()},
\code{\link{col_names}()},
\code{\link{matrix_int_names}()},
\code{\link{row_names}()}
}
\concept{matrix_int_names_functions}
