% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q_matrix.R
\name{q_matrix}
\alias{q_matrix}
\alias{q_matrix_list}
\title{Read and structure .Q files or existing matrices as \code{q_matrix} or \code{q_matrix_list} objects.}
\usage{
q_matrix(x)
}
\arguments{
\item{x}{can be:
\itemize{
\item a path to a directory containing .Q files
\item a path to a single .Q file
\item a matrix
\item a dataframe
\item a list of dataframes or matrices
}}
}
\value{
either:
\itemize{
\item a single \code{q_matrix} object
\item a \code{q_matrix_list} object containing a list of Q matrices and a list of
indices for each Q matrix separated by K
}
}
\description{
This function reads .Q matrix files generated by external clustering
algorithms (such as ADMIXTURE) and transforms them into \code{q_matrix} or
\code{q_matrix_list} objects for plotting.
}
