% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q_matrix.R
\name{q_matrix}
\alias{q_matrix}
\title{Read and structure .Q files or existing matrices as 'q_matrix' objects.}
\usage{
q_matrix(x)
}
\arguments{
\item{x}{can be:
\itemize{
\item a path to a directory containing .Q files
\item a path to a single .Q file
\item a matrix
}}
}
\value{
either:
\itemize{
\item a single \code{q_matrix} object
\item a list of \code{q_matrix} objects
\item a \code{q_matrix_list} object (for directories containing multiple runs) of q_matrix objects
}
}
\description{
This function reads .Q matrix files generated by external clustering
algorithms (such as ADMIXTURE) and transforms them into 'q_matrix' objects
for plotting.
}
