% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indiv_qc_report.R
\name{qc_report_indiv}
\alias{qc_report_indiv}
\title{Create a Quality Control report for individuals}
\usage{
qc_report_indiv(.x, kings_threshold = NULL, ...)
}
\arguments{
\item{.x}{a \code{\link{gen_tibble}} object.}

\item{kings_threshold}{an optional numeric, a threshold of relatedness for the sample}

\item{...}{further arguments to pass}
}
\value{
a tibble with 2 elements: het_obs and missingness
}
\description{
#' Return QC information to assess loci (Observed heterozygosity and missingness).
}
