% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbind_dry_run.R
\name{rbind_dry_run}
\alias{rbind_dry_run}
\title{Generate a report of what would happen to each SNP in a merge}
\usage{
rbind_dry_run(
  ref,
  target,
  flip_strand = FALSE,
  remove_ambiguous = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{ref}{either a \code{\link{gen_tibble}} object, or the path to the PLINK file;
the alleles in this objects will
be used as template to flip the ones in \code{target} and/or
swap their order as necessary.}

\item{target}{either a \code{\link{gen_tibble}} object, or the path to the PLINK file
(saved in raw format, see details in \code{\link[=read_plink_raw]{read_plink_raw()}}.}

\item{flip_strand}{boolean on whether strand flipping should be checked to
match the two datasets. It defaults to FALSE}

\item{remove_ambiguous}{boolean whether ambiguous SNPs (i.e. a/t and c/g)
should be removed. It defaults to FALSE}

\item{quiet}{boolean whether to omit reporting to screen}
}
\value{
a list with two \code{data.frames}, named \code{target} and \code{ref}. Each
data.frame has \code{nrow()} equal to the number of loci in the respective dataset,
a column \code{id} with the locus name, and boolean columns \code{to_keep} (the valid loci
that will be kept in the merge),
\code{alleles_mismatched} (loci found in both datasets but with mismatched alleles,
leading to those loci being dropped), \code{to_flip} (loci that need to be flipped
to align the two datasets, only found in \code{target} data.frame)
and \code{to_swap} (loci for which the order of alleles needs to be swapped
to align the two datasets, \code{target} data.frame)
}
\description{
This function provides an overview of the fate of each SNP in two
\code{\link[adegenet:genlight]{adegenet::genlight}} objects in the case of a merge. Only SNPs found in
both objects will be kept. One object is used as a \code{reference}, and SNPs
in the other dataset will be flipped and/or alleles swapped
as needed. SNPs that have different alleles in the two datasets will also be
dropped.
}
