% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_fill_distruct.R
\name{scale_fill_distruct}
\alias{scale_fill_distruct}
\title{Scale constructor using the distruct colours}
\usage{
scale_fill_distruct(guide = "none", ...)
}
\arguments{
\item{guide}{guide function passed to \code{\link[ggplot2:scale_manual]{ggplot2::scale_fill_manual()}}.
Defaults to "none", set to "legend" if a legend is required.}

\item{...}{further parameters to be passed to \code{\link[ggplot2:scale_manual]{ggplot2::scale_fill_manual()}}}
}
\value{
a scale constructor to be used with ggplot
}
\description{
A wrapper around \code{\link[ggplot2:scale_manual]{ggplot2::scale_fill_manual()}}, using the distruct colours
from \code{\link{distruct_colours}}.
}
