% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_loci_if.R
\name{select_loci_if}
\alias{select_loci_if}
\title{The \code{select_if} verb for \code{loci}}
\usage{
select_loci_if(.data, .sel_logical)
}
\arguments{
\item{.data}{a \code{gen_tibble}}

\item{.sel_logical}{a logical vector of length equal to the number of loci,
or an expression that will tidy evaluate to such a vector}
}
\value{
a list of \code{SNPbin} object that have been subsetted.
}
\description{
An equivalent to \code{\link[dplyr:select_all]{dplyr::select_if()}} that works on the \code{genotype} column
of a \code{gen_tibble}. Tidy evaluation should work as expected.
}
