% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_genotypes.R
\name{show_genotypes}
\alias{show_genotypes}
\alias{show_genotypes.gen_tbl}
\alias{show_genotypes.list}
\title{Show the genotypes of a \code{gen_tibble}}
\usage{
show_genotypes(.data, ...)

\method{show_genotypes}{gen_tbl}(.data, ...)

\method{show_genotypes}{list}(.data, ...)
}
\arguments{
\item{.data}{a \code{\link{gen_tibble}} object.}

\item{...}{currently unused}
}
\value{
a \code{\link[tibble:tibble]{tibble::tibble}} of information (see\code{\link{gen_tibble}} for details
on compulsory columns that will always be present)
}
\description{
Extract the genotypes (as a matrix) from a  \code{gen_tibble}.
}
