% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_genotypes.R
\name{show_genotypes}
\alias{show_genotypes}
\alias{show_genotypes.gen_tbl}
\alias{show_genotypes.list}
\title{Show the genotypes of a \code{gen_tibble}}
\usage{
show_genotypes(.x, ...)

\method{show_genotypes}{gen_tbl}(.x, ...)

\method{show_genotypes}{list}(.x, ...)
}
\arguments{
\item{.x}{a list of \code{\link[adegenet:SNPbin]{adegenet::SNPbin}} objects (usually the \code{genotype} column of
a \code{\link{gen_tibble}} object),
or a \code{\link{gen_tibble}}.}

\item{...}{currently unused.}
}
\value{
a matrix of counts of the alternative alleles (see \code{\link[=show_loci]{show_loci()}}) to
extract information on the alleles for those loci from a \code{\link{gen_tibble}}.
}
\description{
Extract the genotypes (as a matrix) from a  \code{gen_tibble}.
}
