% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_int_names.R
\name{[<-.matrix_int_names}
\alias{[<-.matrix_int_names}
\title{Assignment method for matrix_int_names}
\usage{
\method{[}{matrix_int_names}(x, i, j, i_names = NULL, j_names = NULL) <- value
}
\description{
Assignment method for matrix_int_names
}
