% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_distruct.R
\name{theme_distruct}
\alias{theme_distruct}
\title{A theme to match the output of distruct}
\usage{
theme_distruct()
}
\value{
a \link[ggplot2:theme]{ggplot2::theme}
}
\description{
A theme to remove most plot decorations, matching the look of plots
created with distruct.
}
\examples{
# Read example gt_admix object
admix_obj <-
  readRDS(system.file("extdata", "anolis", "anole_adm_k3.rds",
    package = "tidypopgen"
  ))

# Basic barplot with disstruct theme
autoplot(admix_obj, k = 3, run = 1, type = "barplot") +
  theme_distruct()
}
