% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_q_matrix.R
\name{tidy.q_matrix}
\alias{tidy.q_matrix}
\title{Tidy a Q matrix}
\usage{
\method{tidy}{q_matrix}(x, gen_tbl, ...)
}
\arguments{
\item{x}{A Q matrix object (as returned by LEA::Q()).}

\item{gen_tbl}{An associated gen_tibble}

\item{...}{not currently used}
}
\value{
A tidied tibble
}
\description{
Takes a \code{q_matrix} object, which is a matrix, and returns a tidied tibble.
}
