% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q_matrix.R
\name{tidy.q_matrix}
\alias{tidy.q_matrix}
\title{Tidy a Q matrix}
\usage{
\method{tidy}{q_matrix}(x, data, ...)
}
\arguments{
\item{x}{A Q matrix object (as returned by \link{'q_matrix'}).}

\item{data}{An associated tibble (e.g. a \code{\link{gen_tibble}}), with the individuals in the same order as the data used to
generate the Q matrix}

\item{...}{not currently used}
}
\value{
A tidied tibble
}
\description{
Takes a \code{q_matrix} object, which is a matrix, and returns a tidied tibble.
}
