% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidypopgen-package.R
\docType{package}
\name{tidypopgen-package}
\alias{tidypopgen}
\alias{tidypopgen-package}
\title{tidypopgen: Tidy Population Genetics}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

We provide a tidy grammar of population genetics, facilitating the manipulation and analysis of data on biallelic single nucleotide polymorphisms (SNPs).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://evolecolgroup.github.io/tidypopgen}
}

}
\author{
\strong{Maintainer}: Andrea Manica \email{am315@cam.ac.uk} (\href{https://orcid.org/0000-0003-1895-450X}{ORCID})

Authors:
\itemize{
  \item Evie Carter
}

}
\keyword{internal}
