% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidypopgen-package.R
\docType{package}
\name{tidypopgen-package}
\alias{tidypopgen}
\alias{tidypopgen-package}
\title{tidypopgen: Tidy Population Genetics}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

We provide a tidy grammar of population genetics, facilitating the manipulation and analysis of data on biallelic single nucleotide polymorphisms (SNPs). `tidypopgen` scales to very large genetic datasets by storing genotypes on disk, and performing operations on them in chunks, without ever loading all data in memory.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/EvolEcolGroup/tidypopgen}
  \item \url{https://evolecolgroup.github.io/tidypopgen/}
  \item Report bugs at \url{https://github.com/EvolEcolGroup/tidypopgen/issues}
}

}
\author{
\strong{Maintainer}: Andrea Manica \email{am315@cam.ac.uk} (\href{https://orcid.org/0000-0003-1895-450X}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Evie Carter
  \item Eirlys Tysall
}

}
\keyword{internal}
