% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_to_fbm_cpp_fast.R
\name{vcf_to_fbm_cpp}
\alias{vcf_to_fbm_cpp}
\title{Convert vcf to FBM.}
\usage{
vcf_to_fbm_cpp(vcf_path, backingfile = NULL, quiet = FALSE)
}
\arguments{
\item{vcf_path}{the path to the vcf}

\item{backingfile}{the name of the file to use as the backing file for the
FBM. If NULL, the vcf path will be used.}
}
\value{
path to the resulting rds file as class bigSNP.
}
\description{
Convert a vcf file to a Filebacked Big Matrix (FBM) object. This should work
even for large vcf files that would not fit in memory.
}
\keyword{internal}
