% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windows_pop_tajimas_d.R
\name{windows_pop_tajimas_d}
\alias{windows_pop_tajimas_d}
\title{Compute Tajima's D for a sliding window}
\usage{
windows_pop_tajimas_d(
  .x,
  window_size,
  step_size,
  size_unit = c("snp", "bp"),
  min_loci = 1,
  complete = FALSE
)
}
\arguments{
\item{.x}{a (potentially grouped) \code{gen_tibble} object}

\item{window_size}{The size of the window to use for the estimates.}

\item{step_size}{The step size to use for the windows.}

\item{size_unit}{Either "snp" or "bp". If "snp", the window size and step
size are in number of SNPs. If "bp", the window size and step size are in
base pairs.}

\item{min_loci}{The minimum number of loci required to calculate a window
statistic. If the number of loci in a window is less than this, the window
statistic will be NA.}

\item{complete}{Should the function be evaluated on complete windows only? If
FALSE, the default, then partial computations will be allowed at the end of
the chromosome.}
}
\value{
a data frame (or a list of data.frames if \code{.x} is grouped) with the
following columns:
\itemize{
\item \code{chromosome}: the chromosome for the window
\item \code{start}: the starting locus of the window
\item \code{end}: the ending locus of the window
\item \code{tajimas_d}: the Tajimas'D for the population
}
}
\description{
This function computes Tajimas's D for a sliding window across
each chromosome.
}
